API

Types

FrameTransformations.Frames.AbstractFrameAxes — Type

AbstractFrameAxes

Abstract type for all reference frames axes.

FrameTransformations.Frames.AbstractFramePoint — Type

AbstractFramePoint

Abstract type for all reference frames points.

FrameTransformations.Frames.ComputableAxesProperties — Type

ComputableAxesProperties

Store the properties required to retrieve all the vectors required by a computable set of axes.

FrameTransformations.Frames.ComputableAxesVector — Type

ComputableAxesVector(from, to, order::Int)

Store the properties required to retrieve the i-th order components of a desired vector. Arguments from and to are the NAIFIDs or AbstractFramePoint instances that define the observer and target points.

Only orders between 1 and 3 are supported.

Example

julia> @point SSB 0 SolarSystemBarycenter

julia> @point Sun 10 

julia> ComputableAxesVector(SSB, Sun, 1)
ComputableAxesVector(0, 10, 1)

julia> ComputableAxesVector(0, 10, 1)
ComputableAxesVector(0, 10, 1)

FrameTransformations.Frames.FrameAxesNode — Type

FrameAxesNode{O, T, N} <: AbstractGraphNode

Define a set of axes.

Fields

name – axes name

class – Symbol representing the class of the axes

id – axes ID (equivalent of NAIFId for axes)

parentid – ID of the parent axes

comp – properties for computable axes

R – vector storing rotation matrices

epochs – vector storing the epochs associated to R

nzo – last order at which R has been computed

f – FrameAxesFunctions container

angles – vector storing the libration angles retrived from ephemerides

FrameTransformations.Frames.FramePointNode — Type

FramePointNode{O, T, N} <: AbstractGraphNode

Define a frame system point.

Fields

name – point name

class – Symbol representing the class of the point

axesid – ID of the axes in which the point coordinates are expressed

parentid – NAIF ID of the parent point

NAIFId – NAIF ID of the point

stv – vector storing the point state vectors

epochs – vector storing the epochs associated to stv

nzo – last order at which stv has been computed

f – FramePointFunctions container

FrameTransformations.Frames.FrameSystem — Type

FrameSystem{O, T, S, E}

A FrameSystem instance manages a collection of user-defined FramePointNode and FrameAxesNode objects, enabling efficient computation of arbitrary transformations between them. It is created by specifying the maximum transformation order O, the outputs datatype T and an AbstractTimeScale instance S. Additionally, an AbstractEphemerisProvider instance E can be provided to compute transformations that require ephemeris data.

The following transformation orders are accepted:

1: position

2: position and velocity

3: position, velocity and acceleration

4: position, velocity, acceleration and jerk

By specifying the maximum transformation the FrameSystem memory usage and performance can be optimised and tailored to the user's needs.

FrameSystem{O, T}()

Create a FrameSystem object of order O and datatype T. The BarycentricDynamicalTime is automatically assigned as the default time scale. The resulting object is constructed with a NullEphemerisProvider, which does not allow the computation of transformation that involve ephemeris files.

Examples

julia> F = FrameSystem{2, Float64}();

julia> @axes ICRF 1 

julia> @axes ECLIPJ2000 17 

julia> add_axes_inertial!(F, ICRF)

julia> add_axes_eclipj2000!(F, ECLIPJ2000, ICRF)

julia> rotation6(F, ICRF, ECLIPJ2000, 0.0)
Rotation{2, Float64}
[...]

julia> rotation9(F, ICRF, ECLIPJ2000, 0.0)
ERROR: Insufficient frame system order: transformation requires at least order 3.

FrameSystem{O, T, S}()

Create a FrameSystem object of order O, datatype T and time scale S. The resulting object is constructed with a NullEphemerisProvider, which does not allow the computation of transformation that involve ephemeris files.

Examples

julia> F = FrameSystem{1, Float64, TerrestrialTime}();

julia> @axes ICRF 1 

julia> @axes ECLIPJ2000 17 

julia> add_axes_inertial!(F, ICRF)

julia> add_axes_eclipj2000!(F, ECLIPJ2000, ICRF)

julia> ep_tt = Epoch("2023-02-10T12:00:00 TT")
2023-02-10T12:00:00.000 TT

julia> rotation3(F, ICRF, ECLIPJ2000, ep_tt)
Rotation{1,Float64}([...])

julia> ep_tdb = Epoch("2023-02-10T12:00:00 TDB")
2023-02-10T12:00:00.000 TDB

julia> rotation3(F, ICRF, ECLIPJ2000, ep_tdb)
ERROR: ArgumentError: Incompatible epoch timescale: expected TerrestrialTime, found BarycentricDynamicalTime.
[...]

FrameSystem{O, T}(eph::AbstractEphemerisProvider)

Create a FrameSystem object of order O and datatype T by providing an instance of an AbstractEphemerisProvider subtype. The timescale is automatically set to the one associated to the ephemeris files loaded in eph. This constructor shall be used when the user desires to compute transformations that involve ephemeris data.

Note

All the kernels that will be used must be loaded within eph. Once the FrameSystem has been created, no additional kernel can be added nor removed.

Examples

julia> eph = CalcephProvider(DE440_KERNEL_PATH)
CalcephProvider(CALCEPH.Ephem(Ptr{Nothing} @0x0000000009e210c0))

julia> F = FrameSystem{2, Float64}(eph)
FrameSystem{2, Float64, BarycentricDynamicalTime, CalcephProvider}(
  eph: CalcephProvider(CALCEPH.Ephem(Ptr{Nothing} @0x0000000009e210c0)),
  points: EMPTY
  axes: EMPTY
)

See also

See also add_axes_inertial!, add_point_root!, vector3 and rotation3

FrameTransformations.Frames.Rotation — Type

Rotation{S, N}

A container to efficiently compute S-th order rotation matrices of type N between two set of axes. It stores the Direction Cosine Matrix (DCM) and its time derivatives up to the (S-1)-th order. Since this type is immutable, the data must be provided upon construction and cannot be mutated later.

The rotation of state vector between two set of axes is computed with an ad-hoc overload of the product operator. For example, a 3rd order Rotation object R, constructed from the DCM A and its time derivatives δA and δ²A rotates a vector v = [p, v, a] as:

̂v = [A*p, δA*p + A*v, δ²A*p + 2δA*v + A*a]

A Rotation object R call always be converted to a SMatrix or a MMatrix by invoking the proper constructor.

Examples

julia> A = angle_to_dcm(π/3, :Z)
DCM{Float64}:
  0.5       0.866025  0.0
 -0.866025  0.5       0.0
  0.0       0.0       1.0

julia> R = Rotation(A);

julia> SM = SMatrix(R)
3×3 SMatrix{3, 3, Float64, 9} with indices SOneTo(3)×SOneTo(3):
  0.5       0.866025  0.0
 -0.866025  0.5       0.0
  0.0       0.0       1.0

julia> MM = MMatrix(R)
3×3 MMatrix{3, 3, Float64, 9} with indices SOneTo(3)×SOneTo(3):
  0.5       0.866025  0.0
 -0.866025  0.5       0.0
  0.0       0.0       1.0

Rotation(dcms::DCM...)

Create a Rotation object from a Direction Cosine Matrix (DCM) and any of its time derivatives. The rotation order is inferred from the number of inputs, while the rotation type is obtained by promoting the DCMs types.

Examples

julia> A = angle_to_dcm(π/3, :Z); 

julia> δA = DCM(0.0I); 

julia> δ²A = DCM(0.0I); 

julia> R = Rotation(A, δA, δ²A); 

julia> typeof(R) 
Rotation{3, Float64}

julia> R2 = Rotation(A, B, C, DCM(0.0im*I)); 

julia> typeof(R2)
Rotation{4, ComplexF64}

Rotation{S}(u::UniformScaling{N}) where {S, N}
Rotation{S, N}(u::UniformScaling) where {S, N}

Create an S-order identity Rotation object of type N with identity position rotation and null time derivatives.

Examples

julia> Rotation{1}(1.0I) 
Rotation{1, Float64}(([1.0 0.0 0.0; 0.0 1.0 0.0; 0.0 0.0 1.0],))

julia> Rotation{1, Int64}(I)
Rotation{1, Int64}(([1 0 0; 0 1 0; 0 0 1],))

Rotation{S1}(rot::Rotation{S2, N}) where {S1, S2, N}

Transform a Rotation object of order S2 to order S1. This conversion is only possible if S1 < S2.

Examples

julia> A = angle_to_dcm(π/3, :Z);

julia> B = angle_to_dcm(π/4, π/6, :XY);

julia> R1 = Rotation(A, B);

julia> order(R1)
2

julia> R2 = Rotation{1}(R1);

julia> order(R2)
1

julia> R2[1] == A 
true

julia> R3 = Rotation{3}(R1)
ERROR: DimensionMismatch: Cannot convert a `Rotation` of order 2 to order 3
[...]

Rotation(m::DCM{N}, ω::AbstractVector) where N

Create a 2nd order Rotation object of type N to rotate between two set of axes a and b from a Direction Cosine Matrix (DCM) and the angular velocity vector ω of b with respect to a, expressed in b

Rotation{S}(dcms::DCM...) where S

Create a Rotation object of order S. If the number of dcms is smaller than S, the remaining slots are filled with null DCMs, otherwise if the number of inputs is greater than S, only the first S DCMs are used.

Warning

Usage of this constructor is not recommended as it may yield unexpected results to unexperienced users.

See also

See also rotation3, rotation6 and rotation9.

Functions

Base.inv — Method

inv(rot::Rotation)

Compute the invese of the rotation object rot. The operation is efficiently performed by taking the transpose of each rotation matrix within rot.

FrameTransformations.Frames._get_comp_axes_vector3 — Method

_get_comp_axes_vector3(frame, v, axesid, t)

Compute a 3-elements vector in the desired axes at the given time between two points of the frame system

The returned vector depends on the order in v as follows:

1: position

2: velocity

3: acceleration

FrameTransformations.Frames._two_vectors_basis — Method

_two_vectors_basis(a, b, seq::Symbol, fc::Function)

Generate a 3D right-handed orthogonal vector basis and/or its time derivatives from the vectors a and b, according to the directions specified in seq and the input cross function fc.

The accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

The standard basis, its 1st, 2nd-order and 3rd order time derivatives can be computed by passing cross, cross6, cross9 or cross12 to fc. The returned vectors will have a length of 3, 6 or 9, respectively.

FrameTransformations.Frames._two_vectors_to_rot12 — Method

_two_vectors_to_rot12(a, b, seq::Symbol)

Generate a direction cosine matrix and its 1st, 2nd and 3r-order time-derivatives, minimising the number of repeated computations.

See also

See twovectors_to_dcm and twovectors_to_δdcm for more information.

FrameTransformations.Frames._two_vectors_to_rot9 — Method

_two_vectors_to_rot9(a, b, seq::Symbol)

Generate a direction cosine matrix and its 1st and 2nd-order time-derivatives, minimising the number of repeated computations.

See also

See twovectors_to_dcm and twovectors_to_δdcm for more information.

FrameTransformations.Frames._twovectors_to_dcm — Method

_twovectors_to_dcm(a, b, seq::Symbol, fc::Function, fn::Function)

Generate a direction cosine matrix and/or its time derivatives from the vectors a and b, according to the directions specifeid in seq.

Notes

fc and fn are used to control the derivative order.

FrameTransformations.Frames.add_axes_bci2000! — Method

add_axes_bci2000!(frames::FrameSystem{O, T}, 
    data::AbstractDict, center::AbstractFramePoint, 
    axes::AbstractFrameAxes, parent::AbstractFrameAxes) where {T, O}

Insert Body-Centered Inertial (BCI) axes at J2000 relative to the body center to the frames system.

Input/s

frames - The frame system to which the new frame will be added.

data - A dictionary containing a parsed TPC file. The dictionary as has keys the NAIFId of the bodies contained in the file, and as values a dictionary containing the keywords and the values of the file's data.

center: The center point of the new axes.

axes: The new axes to be added to the frame system.

parent: The parent axes of the new ones.

Warning

The parent set of axes must be the International Celestial Reference Frame (ICRF). If the parent set of axes is not ICRF, an error is thrown.

FrameTransformations.Frames.add_axes_bcrtod! — Method

add_axes_bcrtod!(frames::FrameSystem{O, T}, 
    data::AbstractDict, center::AbstractFramePoint, 
    axes::AbstractFrameAxes, parent::AbstractFrameAxes) where {T, O}

Insert a Body-Centered Rotating (BCR), True-of-Date (TOD) axes to the frames system.

Input/s

frames - The frame system to which the new frame will be added.

data - A dictionary containing a parsed TPC file. The dictionary as has keys the NAIFId of the bodies contained in the file, and as values a dictionary containing the keywords and the values of the file's data.

center: The center point of the new axes.

axes: The new axes to be added to the frame system.

parent: The parent axes of the new ones.

Note

The axes constructed here corresponds to the SPICE IAU_<BODY_NAME> frames.

Warning

The parent set of axes must be the International Celestial Reference Frame (ICRF). If the parent set of axes is not ICRF, an error is thrown.

FrameTransformations.Frames.add_axes_computable! — Method

add_axes_computable!(frames, axes, parent, v1, v2, seq::Symbol)

Add axes as a set of computable axes to frames. Computable axes differ from rotating axes because they are computed through two vectors that are defined within the frame system itself. Computable axes are the equivalent of SPICE's parameterized two-vector frames.

These axes are built such that the first vector, known as the primary vector, is parallel to one axis of the frame; the component of the secondary vector orthogonal to the first is parallel to another axis of the frame, and the cross product of the two vectors is parallel to the remaining axis.

The primary and secondary vectors, v1 and v2 are instances of ComputableAxesVector, which is used to define the NAIF IDs of the vector origin and target, and its order. Current accepted order values are: 1 (position), 2 (velocity) and 3 (acceleration).

For example, to define a primary vector that is parallel to the Sun's (NAIF ID = 10) velocity with respect to the Solary system barycenter (NAIF ID = 0), v1 must be set as: v1 = ComputableAxesVector(10, 0, 1).

seq is a combination of two letters that is used to identify the desired pointing directions of the primary and secondary vectors. Accepted sequences are: :XY, :YX, :XZ, :ZX, :YZ and :ZY.

Given a spacecraft registered as a point in the frame system, an example of a set of computable axes is the Local Vertical Local Horizon (LVLH), where the spacecraf's nadir direction and velocity direction define the axes orientation.

Note

Regardless of the original set of axes in which the primary and secondary vectors are defined, the axes orientation is automatically computed by rotating them to parent.

Examples

julia> eph = CalcephProvider(DE440_KERNEL_PATH);

julia> FRAMES = FrameSystem{4, Float64}(eph);

julia> @point SSB 0 SolarySystemBarycenter 

julia> @point Sun 10 

julia> @axes ICRF 1 InternationalCelestialReferenceFrame 

julia> add_axes_inertial!(FRAMES, ICRF)

julia> add_point_root!(FRAMES, SSB, ICRF)

julia> add_point_ephemeris!(FRAMES, Sun, SSB)

julia> @axes SunFrame 2

julia> v1 = ComputableAxesVector(10, 0, 1)
ComputableAxesVector(10, 0, 1)

julia> v2 = ComputableAxesVector(10, 0, 2)
ComputableAxesVector(10, 0, 2)

julia> add_axes_computable!(FRAMES, SunFrame, ICRF, v1, v2, :XY)

See also

See also ComputableAxesVector, add_axes_fixedoffset!, add_axes_inertial! and add_axes_computable!

FrameTransformations.Frames.add_axes_eclipj2000! — Method

add_axes_eclipj2000!(frames, axes, parent::AbstractFrameAxes, iau_model::IAUModel=iau1980)

Add axes as a set of inertial axes representing the Ecliptic Equinox of J2000 (ECLIPJ2000) to frames. The obliquity of the ecliptic is computed using the IAU Model iau_model.

The admissed parent set of axes are the following:

ICRF: for the International Celestial Reference Frame, with ID = 1

MEME2000: the Mean Earth/Moon Ephemeris of J2000, with ID = 22

Warning

If the name (or the axes ID) of the parent set of axes is neither ICRF (ID = 1) nor MEME2000 (ID = 22), an error is thrown.

Examples

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> @axes ECLIPJ2000 17 EclipticEquinoxJ2000 

julia> FRAMES = FrameSystem{1, Float64}();

julia> add_axes_inertial!(FRAMES, ICRF)

julia> add_axes_eclipj2000!(FRAMES, ECLIPJ2000, ICRF)

See also

See also add_axes_inertial! and Orient.DCM_ICRF_TO_J2000_BIAS

FrameTransformations.Frames.add_axes_ephemeris! — Method

add_axes_ephemeris!(frames, axes, rot_seq::Symbol)
add_axes_ephemeris!(frames, axes, fun, δfun=nothing, δ²fun=nothing, δ³fun=nothing)

Add axes as a set of ephemeris axes to frames. The orientation of these axes is computed with a series of 3 rotations specified by rot_seq. The euler angles and their derivatives needed by the rotation are extracted from the ephemeris kernels loaded in frames. The parent axes are automatically assigned to the axes with respect to which the orientation data has been written in the kernels.

The rotation sequence is defined by a Symbol specifing the rotation axes. This function assigns dcm = A3 * A2 * A1 in which Ai is the DCM related with the i-th rotation. The possible rot_seq values are: :XYX, XYZ, :XZX, :XZY, :YXY, :YXZ, :YZX, :YZY, :ZXY, :ZXZ, :ZYX, or :ZYZ.

Alternatively, one can provide custom functions that take a vector of euler angles and their derivatives and return the DCM and its derivatives.

This operation is only possible if the ephemeris kernels loaded within frames contain orientation data for the AXES ID associated to axes. An error is returned if the parent axes ID is yet to be added to frames.

Warning

It is expected that the axes ID assigned by the user are aligned with those used to generate the ephemeris kernels. No check are performed on whether these IDs represent the same physical axes that are intended in the kernels.

See also

See also ComputableAxesVector, add_axes_fixedoffset!, add_axes_inertial! and add_axes_computable!

FrameTransformations.Frames.add_axes_fixedoffset! — Method

add_axes_fixedoffset!(frames::FrameSystem{T}, axes, parent, dcm::DCM{T}) where T

Add axes as a set of fixed offset axes to frames. Fixed offset axes have a constant orientation with respect to their parent axes, represented by dcm, a Direction Cosine Matrix (DCM).

add_axes_fixedoffset!(frames, name::Symbol, axesid::Int, parentid::Int, dcm:DMC{T}) where {T}

Add new axes name with id axesid to frames with a fixed-offset from parentid. Low-level function.

Note

While inertial axes do not rotate with respect to the star background, fixed offset axes are only constant with respect to their parent axes, but might be rotating with respect to some other inertial axes.

Examples

julia> FRAMES = FrameSystem{1, Float64}();

julia> @axes ICRF 1 InternationalCelestialReferenceFrame 

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @axes ECLIPJ2000 17 

julia> add_axes_fixedoffset!(FRAMES, ECLIPJ2000, ICRF, angle_to_dcm(π/3, :Z))

See also

See also add_axes_rotating!, add_axes_inertial! and add_axes_computable!

FrameTransformations.Frames.add_axes_inertial! — Method

add_axes_inertial!(frames, axes; parent=nothing, dcm=nothing)

Add axes as a set of inertial axes to frames. Only inertial axes can be used as root axes to initialise the axes graph. Only after the addition of a set of inertial axes, other axes classes may be added aswell. Once a set of root-axes has been added, parent and dcm become mandatory fields.

add_axes_inertial!(frames, name::Symbol, axesid::Int; parentid=nothing, dcm=nothing)

Add new axes name with id axesid to frames. Low-level function.

Note

The parent of a set of inertial axes must also be inertial. FixedOffset axes that are that only have inertial parents are also accepted.

Examples

julia> FRAMES = FrameSystem{2, Float64}();

julia> @axes ICRF 1 InternationalCelestialReferenceFrame 

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @axes ECLIPJ2000 17 

julia> add_axes_inertial!(FRAMES, ECLIPJ2000)
ERROR: A set of parent axes for ECLIPJ2000 is required
[...]

julia> add_axes_inertial!(FRAMES, ECLIPJ2000; parent=ICRF, dcm=angle_to_dcm(π/3, :Z))

See also

See also add_axes_rotating!, add_axes_fixedoffset! and add_axes_computable!

FrameTransformations.Frames.add_axes_itrf! — Method

add_axes_itrf!(frames, axes, parent, model::Orient.IAU2006Model=Orient.iau2006b)

Add axes as a set of axes representing the International Terrestrial Reference Frame (ITRF) to frames. Only the :ICRF and :GCRF are accepted as parent axes. Use the model argument to specify which IAU model model should be used for the computations. The default is set to iau2006b.

See also

See also orient_rot3_itrf_to_gcrf

FrameTransformations.Frames.add_axes_me421! — Method

add_axes_me421!(frames, axes, parent)

Add axes as fixed offset axes representing the DE421 Moon's Mean Earth/Mean Rotation (ME) to frames.

The parent set of axes can be either the DE440 Principal Axes (PA440) or the DE421 Principal Axes (PA421), otherwise an error is thrown. Depending on that, the relative axes orientation will be automatically selected by this function.

See also

See also [`add_axes_pa440!`](@ref), and [`add_axes_pa421!`](@ref), 
[`Orient.DCM_MOON_PA421_TO_ME421`](@ref) and [`Orient.DCM_MOON_PA421_TO_ME421`](@ref),

FrameTransformations.Frames.add_axes_meme2000! — Method

add_axes_meme2000!(frames::, axes, parent::AbstractFrameAxes)

Add axes as a set of inertial axes representing the Mean Equator Mean Equinox of J2000 to frames.

Warning

The name (or the axes ID) of the parent set of axes must be ICRF (i.e., the International Celestial Reference Frame, ID = 1), or the ECLIPJ2000 (i.e., the Ecliptic Equinox of 7 J2000, ID = 17), otherwise and error is thrown.

Examples

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> @axes MEME2000 22 MeanEquatorMeanEquinoxJ2000 

julia> FRAMES = FrameSystem{1, Float64}();

julia> add_axes_inertial!(FRAMES, ICRF)

julia> add_axes_meme2000!(FRAMES, MEME2000, ICRF)

See also

See also add_axes_inertial! and Orient.DCM_ICRF_TO_J2000_BIAS

FrameTransformations.Frames.add_axes_mememod! — Method

add_axes_mememod!(frames, axes, parent::AbstractFrameAxes, model::IAU2006Model=iau2006b)

Add axes as a set of projected axes representing the Mean of Date Ecliptic Equinox to frames.

Note

Despite this class of axes has a rotation matrix that depends on time, its derivatives are assumed null.

Warning

The name of the parent set of axes must be the ICRF or have ID = 1 (i.e., the International Celestial Reference Frame), otherwise an error is thrown.

FrameTransformations.Frames.add_axes_pa421! — Method

add_axes_pa421!(frames, axes, parent)

Add axes as a set of ephemeris axes representing the DE421 Moon's Principal Axes (PA) to frames. The libration angles are extracted from the ephemeris kernels loaded within frames, an error is thrown if such orientation data is not available.

The parent set of axes must be named ICRF or have ID = 1 (i.e., the International Celestial Reference Frame), otherwise and error is thrown.

Warning

To properly read the ephemeris kernels, the ID associated to the input axes must match NAIF's FRAME ID for the Moon PA DE421 axes (31006).

See also

See also [`Orient.AXESID_MOONPA_DE421`](@ref), [`Orient.orient_icrf_to_pa421`](@ref), 
[`add_axes_pa440!`](@ref), and [`add_axes_me421!`](@ref)

FrameTransformations.Frames.add_axes_pa440! — Method

add_axes_pa440!(frames, axes, parent)

Add axes as a set of ephemeris axes representing the DE440 Moon's Principal Axes (PA) to frames. The libration angles are extracted from the ephemeris kernels loaded within frames, an error is thrown if such orientation data is not available.

The parent set of axes must be named ICRF or have ID = 1 (i.e., the International Celestial Reference Frame), otherwise and error is thrown.

Warning

To properly read the ephemeris kernels, the ID associated to the input axes must match NAIF's FRAME ID for the Moon PA DE440 axes (31008).

See also

See also Orient.AXESID_MOONPA_DE440, Orient.orient_rot3_icrf_to_pa440, add_axes_pa421!, and add_axes_me421!

FrameTransformations.Frames.add_axes_projected! — Method

add_axes_projected!(frames, axes, parent, fun)

Add axes as a set of projected axes to frames. The orientation of these axes depends only on time and is computed through the custom functions provided by the user.

Projected axes are similar to rotating axis, except that all the positions, velocity, etc ... are rotated by the 0-order rotation (i.e. the derivatives of the rotation matrix are null, despite the rotation depends on time).

Warning

It is expected that the input function and their outputs have the correct signature. This function does not perform any checks on the output types.

FrameTransformations.Frames.add_axes_rotating! — Method

add_axes_rotating!(frames, axes, parent, fun, δfun=nothing, δ²fun=nothing, δ³fun=nothing) where T

Add axes as a set of rotating axes to frames. The orientation of these axes depends only on time and is computed through the custom functions provided by the user.

The input functions must accept only time as argument and their outputs must be as follows:

fun: return a Direction Cosine Matrix (DCM).

δfun: return the DCM and its 1st order time derivative.

δ²fun: return the DCM and its 1st and 2nd order time derivatives

δ³fun: return the DCM and its 1st, 2nd and 3rd order time derivatives

If δfun, δ²fun or δ³fun are not provided, they are computed via automatic differentiation.

Warning

It is expected that the input functions and their outputs have the correct signature. This function does not perform any checks on the output types.

Examples

julia> FRAMES = FrameSystem{3, Float64}();

julia> @axes Inertial 1

julia> add_axes_inertial!(FRAMES, Inertial)

julia> @axes Synodic 2 

julia> fun(t) = angle_to_dcm(t, :Z);

julia> add_axes_rotating!(FRAMES, Synodic, Inertial, fun)

julia> R = rotation6(FRAMES, Inertial, Synodic, π/6);

julia> R[1]
DCM{Float64}:
  0.866025  0.5       0.0
 -0.5       0.866025  0.0
  0.0       0.0       1.0

julia> R[2]
DCM{Float64}:
 -0.5        0.866025  0.0
 -0.866025  -0.5       0.0
  0.0        0.0       0.0

See also

See also add_axes_fixedoffset!, add_axes_inertial! and add_axes_computable!

FrameTransformations.Frames.add_axes_topocentric! — Method

add_axes_topocentric!(frames, axes, λ::Number, ϕ::Number, type::Symbol, parent)

Add axes as a set of fixed-offset topocentric axes to frames. The orientation relative to the parent axes parent is defined throuh the longitude λ, the geodetic latitude ϕ and the type type, which may be any of the following:

:NED (North, East, Down): the X-axis points North, the Y-axis is directed eastward and the Z-axis points inwards towards the nadir.

:SEZ (South, East, Zenith): the X-axis points South, the Y-axis is directed East, and the Z-axis points outwards towards the zenith.

:ENU (East, North, Up): the X-axis points East, the Y-axis is directed North and the Z-axis points outwards towards the zenith.

Warning

The parent axes must be a set of body-fixed reference axes. When this constraint is not satisfied, the results may be fundamentally wrong.

See also

See also add_axes_fixedoffset! and add_point_surface!.

FrameTransformations.Frames.add_point_dynamical! — Method

add_point_dynamical!(frames, point, parent, axes, fun, δfun=nothing, δ²fun=nothing, δ³fun=nothing)

Add point as a time point to frames. The state vector for these points depends only on time and is computed through the custom functions provided by the user.

The input functions must accept only time as argument and their outputs must be as follows:

fun: return a 3-elements vector: position

δfun: return a 6-elements vector: position and velocity

δ²fun: return a 9-elements vector: position, velocity and acceleration

δ³fun: return a 12-elements vector: position, velocity, acceleration and jerk

If δfun, δ²fun or δ³fun are not provided, they are computed with automatic differentiation.

Warning

It is expected that the input functions and their ouputs have the correct signature. This function does not perform any checks on whether the returned vectors have the appropriate dimensions.

Examples

julia> FRAMES = FrameSystem{2, Float64}()

julia> @axes ICRF 1 

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point Origin 0 

julia> add_point_root!(FRAMES, Origin, ICRF)

julia> @point Satellite 1 

julia> satellite_pos(t::T) where T = [cos(t), sin(t), 0]

julia> add_point_dynamical!(FRAMES, Satellite, Origin, ICRF, satellite_pos)

julia> vector6(FRAMES, Origin, Satellite, ICRF, π/6)
6-element SVector{6, Float64} with indices SOneTo(6):
  0.8660254037844387
  0.49999999999999994
  0.0
 -0.49999999999999994
  0.8660254037844387
  0.0

See also

See also add_point_root!, add_point_ephemeris!,add_point_fixed! and add_point_updatable!

FrameTransformations.Frames.add_point_ephemeris! — Method

add_point_ephemeris!(frames, point, parent=nothing)

Add point as an ephemeris point to frames. This function is intended for points whose state-vector is read from ephemeris kernels (i.e., de440.bsp). If a parent point is not specified, it will automatically be assigned to the point with respect to which the ephemeris data is written in the kernels.

Inputs

frames – FrameSystem object

point – FramePointNode instance

parent – FramePointNode parent instance

add_point_ephemeris!(frames, name::Symbol, naifid::Int, parentid::Int=nothing, axes=nothing)

Add a new point called name and with id naifid to frames. If a parentid is not specified, it will be automatically assigned depending on the ephemeris kernels.

Inputs

frames – FrameSystem object

name::Symbol – name of the point

naifid::Int – NAIFId associated to the point

parentid::Int – NAIFId associated to the parent point, default is nothing

axes – Id associated to the axes in which the states are expressed, default nothing (taken automatically from the ephemeris files).

Ephemeris points only accept as parent points root-points or other ephemeris points.

Notes

This operation is only possible if the ephemeris kernels loaded within frames contain data for the NAIF ID associated to point and to its parent.

The axes in which the state-vector is expressed are taken from the ephemeris data: an error is returned if the axes ID is yet to be added to frames.

Warning

It is expected that the NAIF ID and the axes ID assigned by the user are aligned with those used to generate the ephemeris kernels. No check are performed on whether these IDs represent the same physical bodies and axes that are intended in the kernels.

Examples

julia> eph = CalcephProvider(DE440_KERNEL_PATH);

julia> FRAMES = FrameSystem{2, Float64}(eph);

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point SSB 0 SolarSystemBarycenter

julia> @point Sun 10 

julia> add_point_root!(FRAMES, SSB, ICRF)

julia> add_point_ephemeris!(FRAMES, Sun)

julia> @point Jupiter 599

julia> add_point_ephemeris!(FRAMES, Jupiter)
ERROR: ArgumentError: Ephemeris data for NAIFID 599 is not available in the kernels loaded in the given FrameSystem.
[...]

See also

See also add_point_root!, add_point_fixed!, add_point_dynamical! and add_point_updatable!

FrameTransformations.Frames.add_point_fixed! — Method

add_point_fixed!(frames, point, parent, axes, offset::AbstractVector)

Add point as a fixed point to frames. Fixed points are those whose positions have a constant offset with respect their parent points in the given set of axes. Thus, points eligible for this class must have null velocity and acceleration with respect to parent.

Inputs

frames – FrameSystem object

point – FramePointNode instance

parent – FramePointNode parent instance

axes – FrameAxesNode instance

offset – Offset with respect to the parent

add_point_fixed!(frames, name::Symbol, pointid::Int, parentid::Int, axes, offset::AbstractVector)

Add a new fixed point called name with id pointid and parent parentid in the axes axesid to frames.

frames – FrameSystem object

name::Symbol – name of the point

naifid::Int – NAIFId associated to the point

parentid::Int – NAIFId associated to the parent point, default is nothing

axesid::Int – Id associated to the axes in which the states are expressed, default nothing (taken automatically from the ephemeris files).

offset – Offset with respect to the parent

Examples

julia> FRAMES = FrameSystem{2, Float64}();

julia> @axes SF -3000 SatelliteFrame

julia> add_axes_inertial!(FRAMES, SF)

julia> @point SC -10000 Spacecraft

julia> @point SolarArrayCenter -10001

julia> add_point_root!(FRAMES, SC, SF)

julia> sa_offset = [0.10, 0.15, 0.30];

julia> add_point_fixed!(FRAMES, SolarArrayCenter, SC, SF, sa_offset)

See also

See also add_point_root!, add_point_ephemeris!, add_point_dynamical! and add_point_updatable!

FrameTransformations.Frames.add_point_root! — Method

add_point_root!(frames, point, axes)

Add point as a root point to frames to initialize the points graph. Only after the addition of a root point, other points may be added aswell. This point is intended as the origin, i.e., its position will equal (0., 0., 0.).

Inputs

frames – FrameSystem object

point – Target point instance

axes – Id or instance of the axes where the point state-vector is expressed.

add_point_root!(frames, name, id, axes)

Add a root point fo frames to initialize the points graphs.

Inputs

frames – FrameSystem object

name – Name of the root point

id – Id (NAIFId) of the root point

axes – Id or instance of the axes where the point state-vector is expressed.

Note

This operation can be performed only once per FrameSystem object: multiple root points in the same graph are both inadmissible and meaningless.

Examples

julia> FRAMES = FrameSystem{2, Float64}();

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point SSB 0 SolarSystemBarycenter 

julia> add_point_root!(FRAMES, SSB, ICRF)

julia> @point Sun 10

julia> add_point_root!(FRAMES, Sun, ICRF)
ERROR: ArgumentError: A root-point is already registed in the given FrameSystem.
[...]

See also

See also add_point_ephemeris!, add_point_fixed!, add_point_dynamical! and add_point_updatable!

FrameTransformations.Frames.add_point_surface! — Function

add_point_surface!(frames, point, parent, axes, pck, λ, ϕ, h::Number=0.0)

Add point to frames as a fixed point on the surface of the parent point. The relative position is specified by the longitude λ, the geodetic latitude ϕ and the altitude over the surface of the reference ellipsoid h, which defaults to 0.0. The ellipsoid parameters are extracted from the input TPC kernel pck using the NAIFId associated to the parent point.

Warning

axes must be a set of body-fixed reference axes for the body represented by parent. When this constraint is not satisfied, the results may be fundamentally wrong.

See also

See also add_point_fixed!, add_axes_topocentric! and geod2pos.

FrameTransformations.Frames.add_point_surface! — Function

add_point_surface!(frames, point, parent, axes, λ, ϕ, R, f::Number=0.0, h::Number=0.0)

Add point to frames as a fixed point on the surface of the parent point body. The relative position is specified by the longitude λ, the geodetic latitude ϕ, the reference radius of the ellipsoid R and its flattening f. The altitude over the reference surface of the ellipsoid h defaults to 0.

FrameTransformations.Frames.add_point_updatable! — Method

add_point_updatable!(frames, point, parent, axes)

Add point as an updatable point to frames. Differently from all the other classes, the state vector for updatable points (expressed in the set of input axes) must be manually updated before being used for other computations.

add_point_updatable!(frames, name::Symbol, pointid::Int, parentid::Int, axes)

Add a point called name with id pointid to frames as a state vector with respect to parentid in the axes axes.

Note

This class of points becomes particularly useful if the state vector is not known a-priori, e.g., when it is the output of an optimisation process which exploits the frame system.

Examples

julia> FRAMES = FrameSystem{2, Float64}();

julia> @axes ICRF 1  

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point Origin 0

julia> @point Satellite 1 

julia> add_point_root!(FRAMES, Origin, ICRF)

julia> add_point_updatable!(FRAMES, Satellite, Origin, ICRF)

julia> y = [10000., 200., 300.]

julia> update_point!(FRAMES, Satellite, y, 0.1)

julia> vector3(FRAMES, Origin, Satellite, ICRF, 0.1)
3-element SVector{3, Float64} with indices SOneTo(3):
 10000.0
   200.0
   300.0

julia> vector3(FRAMES, Origin, Satellite, ICRF, 0.2)
ERROR: UpdatablePoint with NAIFId = 1 has not been updated at time 0.2 for order 1

julia> vector6(FRAMES, Origin, Satellite, ICRF, 0.1)
ERROR: UpdatablePoint with NAIFId = 1 has not been updated at time 0.1 for order 2
[...]

See also

See also update_point!, add_point_root!, add_point_ephemeris!, add_point_dynamical! and add_point_fixed!

FrameTransformations.Frames.axes_alias — Method

axes_alias(ax::AbstractFrameAxes)

Return the axes ID.

FrameTransformations.Frames.axes_id — Function

axes_id(axes::AbstractFrameAxes)

Return the ID associated to axes.

FrameTransformations.Frames.axes_name — Function

axes_name(axes::AbstractFrameAxes)

Return the name of axes.

FrameTransformations.Frames.build_axes — Method

build_axes(frames, name, id, class, funs; parentid, dcm, cax_prop)

Create and add a FrameAxesNode to frames based on the input parameters. Current supported classes are: :InertialAxes, :FixedOffsetAxes, :RotatingAxes, :ProjectedAxes, :EphemerisAxes and :ComputableAxes.

Inputs

frames – Target frame system

name – Axes name, must be unique within frames

id – Axes ID, must be unique within frames

class – Axes class.

funs – FrameAxesFunctions object storing the functions to compute the DCM and, eventually, its time derivatives. It must match the type and order of frames.

Keywords

parentid – Axes ID of the parent axes. Not required only for the root axes.

dcm – DCM with respect to the parent axes. Required only for FixedOffsetAxes.

cax_prop – ComputableAxesProperties, required only by ComputableAxes.

Warning

This is a low-level function and is NOT meant to be directly used. Instead, to add a set of axes to the frame system, see add_axes_inertial!, add_axes_rotating!, add_axes_computable!, add_axes_fixedoffset! and add_axes_projected!.

FrameTransformations.Frames.build_point — Method

build_point(frames, name, NAIFId, class, axesid, funs; parentid, offset)

Create and add a FramePointNode to frames based on the input parameters. Current supported point classes are: :RootPoint, :TimePoint, :EphemerisPoint, :FixedPoint and :UpdatablePoint.

Inputs

frames – Target frame system

name – Point name, must be unique within frames

NAIFId – Point NAIF ID, must be unique within frames

class – Point class.

axesid – ID of the axes in which the state vector of the point is expressed.

funs – FramePointFunctions object storing the functions to update the state vectors of the point. It must match the type and order of frames

Keywords

parentid – NAIF ID of the parent point. Not required only for the root point.

offset – Position offset with respect to a parent point. Required only for FixedPoints.

Warning

This is a low-level function and is NOT meant to be directly used. Instead, to add a point to the frame system, see add_point_ephemeris!, add_point_fixed!, etc...

FrameTransformations.Frames.is_inertial — Method

is_inertial(frame::FrameSystem, axes::AbstractFrameAxes)
is_inertial(frame::FrameSystem, axesid::Int)

Return true if the given axes are inertial, i.e., non rotating with respect to the root inertial axes.

Note

FixedOffsetAxes with respect to an inertial set of axes, are also consired inertial.

FrameTransformations.Frames.is_timefixed — Method

is_timefixed(frame::FrameSystem, axes::AbstractFrameAxes)
is_timefixed(frame::FrameSystem, axesid::Int)

Return true if the given axes are time-fixed, i.e., their orientation does not change in time with respect to the root inertial axes.

Note

Only :InertialAxes and :FixedOffsetAxes defined with respect to other inertial axes are here considered as time fixed.

FrameTransformations.Frames.order — Method

order(R::Rotation{S}) where S

Return the rotation order S.

FrameTransformations.Frames.point_alias — Method

point_alias(ax::AbstractFramePoint)

Return the NAIF ID associated to the input point.

FrameTransformations.Frames.point_id — Function

point_id(point::AbstractFramePoint)

Return the NAIF ID associated to point.

FrameTransformations.Frames.point_name — Function

point_name(point::AbstractFramePoint)

Return the name of point.

FrameTransformations.Frames.rotation12 — Method

rotation12(frame::FrameSystem, from, to, ep::Epoch)

Compute the rotation that transforms a 12-elements state vector from one specified set of axes to another at a given epoch.

Requires a frame system of order ≥ 4.

Inputs

frame – The FrameSystem container object

from – ID or instance of the axes to transform from

to – ID or instance of the axes to transform to

ep – Epoch of the rotation. Its timescale must match that of the frame system.

Output

A Rotation object of order 4.

FrameTransformations.Frames.rotation12 — Method

rotation12(frame::FrameSystem, from, to, t::Number)

Compute the rotation that transforms a 12-elements state vector from one specified set of axes to another at a given time t, expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.rotation3 — Method

rotation3(frame::FrameSystem, from, to, ep::Epoch)

Compute the rotation that transforms a 3-elements state vector from one specified set of axes to another at a given epoch.

Requires a frame system of order ≥ 1.

Inputs

frame – The FrameSystem container object

from – ID or instance of the axes to transform from

to – ID or instance of the axes to transform to

ep – Epoch of the rotation. Its timescale must match that of the frame system.

Output

A Rotation object of order 1.

FrameTransformations.Frames.rotation3 — Method

rotation3(frame::FrameSystem, from, to, t::Number)

Compute the rotation that transforms a 3-elements state vector from one specified set of axes to another at a given time t, expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.rotation6 — Method

rotation6(frame::FrameSystem, from, to, ep::Epoch)

Compute the rotation that transforms a 6-elements state vector from one specified set of axes to another at a given epoch.

Requires a frame system of order ≥ 2.

Inputs

frame – The FrameSystem container object

from – ID or instance of the axes to transform from

to – ID or instance of the axes to transform to

ep – Epoch of the rotation. Its timescale must match that of the frame system.

Output

A Rotation object of order 2.

FrameTransformations.Frames.rotation6 — Method

rotation6(frame::FrameSystem, from, to, t::Number)

Compute the rotation that transforms a 6-elements state vector from one specified set of axes to another at a given time t, expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.rotation9 — Method

rotation9(frame::FrameSystem, from, to, ep::Epoch)

Compute the rotation that transforms a 9-elements state vector from one specified set of axes to another at a given epoch.

Requires a frame system of order ≥ 3.

Inputs

frame – The FrameSystem container object

from – ID or instance of the axes to transform from

to – ID or instance of the axes to transform to

ep – Epoch of the rotation. Its timescale must match that of the frame system.

Output

A Rotation object of order 3.

FrameTransformations.Frames.rotation9 — Method

rotation9(frame::FrameSystem, from, to, t::Number)

Compute the rotation that transforms a 9-elements state vector from one specified set of axes to another at a given time t, expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.twovectors_to_dcm — Method

twovectors_to_dcm(a, b, seq)

Generate a direction cosine matrix from two time-dependent vectors a and b, following the directions specified in seq.

Inputs

a – The primary vector that will be aligned with the first directions specified in seq.

b – The secondary vector. The component of this vector that is orthogonal to the primary vector is aligned with the second direction specified in the sequence seq.

seq – Accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

Warning

The primary and secondary vectors do not have to be orthogonal. However, a great loss of precision happens when the two vectors are almost aligned. This function does not perform any check on the angular separation of the two vectors. The user should ensure that the primary and secondary vector differ of at least 1 milliradian.

FrameTransformations.Frames.twovectors_to_δdcm — Method

twovectors_to_δdcm(a, b, seq)

Compute the time derivative of a direction cosine matrix generated from two time-dependent state vectors a and b, following the directions specified in seq.

Inputs

a and b – 6-elements state vectors (position and velocity).

seq – Accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

FrameTransformations.Frames.twovectors_to_δ²dcm — Method

twovectors_to_δ²dcm(a, b, seq)

Compute the 2nd-order time derivative of a direction cosine matrix generated from two time-dependent state vectors a and b, following the directions specified in seq.

Inputs

a and b – 9-elements state vectors (position velocity and acceleration).

seq – Accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

FrameTransformations.Frames.twovectors_to_δ³dcm — Method

twovectors_to_δ³dcm(a, b, seq)

Compute the 3rd-order time derivative of a direction cosine matrix generated from two time-dependent state vectors a and b, following the directions specified in seq.

Inputs

a and b – 12-elements state vectors (position, velocity, acceleration and jerk).

seq – Accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

FrameTransformations.Frames.update_point! — Method

update_point!(frames, point, stv::AbstractVector, epoch::Epoch)

FrameTransformations.Frames.update_point! — Method

update_point!(frames::FrameSystem, point, stv::AbstractVector, time)

Update the state vector of point at the input time in frames. The only accepted length for the input vector stv are 3, 6, 9 or 12. The order is automatically inferred from the vector length.

Examples

julia> FRAMES = FrameSystem{2, Float64}();
  
julia> @axes ICRF 1  

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point Origin 0

julia> @point Satellite 1 

julia> add_point_root!(FRAMES, Origin, ICRF)

julia> add_point_updatable!(FRAMES, Satellite, Origin, ICRF)

julia> y = [10000., 200., 300.];

julia> update_point!(FRAMES, Satellite, y, 0.1)

julia> vector3(FRAMES, Origin, Satellite, ICRF, 0.1)
3-element SVector{3, StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
 10000.0
   200.0
   300.0

julia> vector3(FRAMES, Origin, Satellite, ICRF, 0.2)
ERROR: UpdatablePoint with NAIFId = 1 has not been updated at time 0.2 for order 1 
[...]

julia> vector6(FRAMES, Origin, Satellite, ICRF, 0.1)
ERROR: UpdatablePoint with NAIFId = 1 has not been updated at time 0.1 for order 2 
[...] 

See also

See also add_point_updatable!

FrameTransformations.Frames.vector12 — Method

vector12(frame::FrameSystem, from, to, axes, ep::Epoch)

Compute 12-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep.

Requires a frame system of order ≥ 4.

Inputs

frame – The FrameSystem container object

from – ID or instance of the observing point

to – ID or instance of the target point

axes – ID or instance of the output state vector axes

ep – Epoch of the observer. Its timescale must match that of the frame system.

FrameTransformations.Frames.vector12 — Method

vector12(frame, from, to, axes, t::Number)

Compute 12-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.vector3 — Method

vector3(frame::FrameSystem, from, to, axes, ep::Epoch)

Compute 3-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep.

Requires a frame system of order ≥ 1.

Inputs

frame – The FrameSystem container object

from – ID or instance of the observing point

to – ID or instance of the target point

axes – ID or instance of the output state vector axes

ep – Epoch of the observer. Its timescale must match that of the frame system.

FrameTransformations.Frames.vector3 — Method

vector3(frame, from, to, axes, ep::Epoch, ltcorr, dir; <keyword arguments>)

Compute a light-time corrected 3-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep, using the aberration flag ltcorr, which may be any of the following AbstractLightTimeCorrection:

LightTimeCorrection: it applies the one-way light time (planetary aberration) correction, using a Newtonian formulation.

PlanetaryAberrationCorrection: it applies the one-way light time and stellar aberration corrections using a Newtonian fromulation. It modifies the vector obtained with the LightTimeCorrection option to account for the observer velocity with respect to the Solar System Barycenter.

The integer argument dir is used to specify the correction direction, as follows:

-1: for Reception, in which photons depart from the target's location at the light-time corrected epoch ep-lt and arrive at the observer's location at ep.

+1: for Transmission, in which photons depart from the observer's location at ep and arrive at the target's location at the light-time corrected epoch ep+lt.

Keyword Arguments

iters::Int=1: the number of iterations used to find the solution to the light time equation. For the solar system bodies, a solution is usually found in 3 iterations.

axescenter: ID or instance of the center point for axes. This parameter is used only when axes have an orientation that depends on time. In these cases, the point is used to find the time at which the orientation of the axes should be computed. It defaults to from.

Note

If the PlanetaryAberrationCorrection is applied, the frame system must be at least one order higher than that of the requested transformation.

See also

See also LightTime, PlanetaryAberration and vector6.

References

CSPICE Library

FrameTransformations.Frames.vector3 — Method

vector3(frame, from, to, axes, t::Number, ltcorr, dir; <keyword arguments>)

Compute a light-time corrected 3-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t, expressed in seconds since J2000, using the aberration flag ltcorr and the direction dir.

FrameTransformations.Frames.vector3 — Method

vector3(frame, from, to, axes, t::Number)

Compute 3-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.vector6 — Method

vector6(frame::FrameSystem, from, to, axes, ep::Epoch)

Compute 6-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep.

Requires a frame system of order ≥ 2.

Inputs

frame – The FrameSystem container object

from – ID or instance of the observing point

to – ID or instance of the target point

axes – ID or instance of the output state vector axes

ep – Epoch of the observer. Its timescale must match that of the frame system.

FrameTransformations.Frames.vector6 — Method

vector6(frame, from, to, axes, ep::Epoch, ltcorr, dir; <keyword arguments>)

Compute a light-time corrected 6-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep, using the aberration flag ltcorr, which may be any of the following AbstractLightTimeCorrection:

LightTimeCorrection: it applies the one-way light time (planetary aberration) correction, using a Newtonian formulation.

PlanetaryAberrationCorrection: it applies the one-way light time and stellar aberration corrections using a Newtonian fromulation. It modifies the vector obtained with the LightTimeCorrection option to account for the observer velocity with respect to the Solar System Barycenter.

The integer argument dir is used to specify the correction direction, as follows:

-1: for Reception, in which photons depart from the target's location at the light-time corrected epoch ep-lt and arrive at the observer's location at ep.

+1: for Transmission, in which photons depart from the observer's location at ep and arrive at the target's location at the light-time corrected epoch ep+lt.

Keyword Arguments

iters::Int=1: the number of iterations used to find the solution to the light time equation. For the solar system bodies, a solution is usually found in 3 iterations.

axescenter: ID or instance of the center point for axes. This parameter is used only when axes have an orientation that depends on time. In these cases, the point is used to find the time at which the orientation of the axes should be computed. It defaults to from.

Note

If the PlanetaryAberrationCorrection is applied, the frame system must be at least one order higher than that of the requested transformation.

See also

See also LightTime, PlanetaryAberration and vector6.

References

CSPICE Library

FrameTransformations.Frames.vector6 — Method

vector6(frame, from, to, axes, t::Number, ltcorr, dir; <keyword arguments>)

Compute a light-time corrected 6-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t, expressed in seconds since J2000, using the aberration flag ltcorr and the direction dir.

FrameTransformations.Frames.vector6 — Method

vector6(frame, from, to, axes, t::Number)

Compute 6-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.vector9 — Method

vector9(frame::FrameSystem, from, to, axes, ep::Epoch)

Compute 9-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep.

Requires a frame system of order ≥ 3.

Inputs

frame – The FrameSystem container object

from – ID or instance of the observing point

to – ID or instance of the target point

axes – ID or instance of the output state vector axes

ep – Epoch of the observer. Its timescale must match that of the frame system.

FrameTransformations.Frames.vector9 — Method

vector9(frame, from, to, axes, t::Number)

Compute 9-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t expressed in seconds since J2000 if ephemerides are used.

Macros

FrameTransformations.Frames.@axes — Macro

@axes(name, id, type=nothing)

Define a new axes instance to alias the given id. This macro creates an AbstractFrameAxes subtype and its singleton instance callen name. Its type name is obtained by appending Axes to either name or type (if provided).

Examples

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> typeof(ICRF)
InternationalCelestialReferenceFrameAxes

julia> axes_alias(ICRF) 
1

julia> @axes IAU_EARTH 10013

julia> typeof(IAU_EARTH)
IauEarthAxes

See also

See also @point and axes_alias.

FrameTransformations.Frames.@point — Macro

@point(name, id, type=nothing)

Define a new point instance to alias the given NAIFID id. This macro creates an AbstractFramePoint subtype and its singleton instance called name. Its type name is obtained by appending Point to either name or type (if provided).

Examples

julia> @point Venus 299

julia> typeof(Venus)
VenusPoint 

julia> point_alias(Venus)
299

julia> @point EMB 3 EarthMoonBarycenter

julia> typeof(EMB) 
EarthMoonBarycenterPoint

julia> point_alias(EMB) 
3 

See also

See also @axes and point_alias.

Constants

FrameTransformations.Frames.LightTime — Constant

LightTime

The singleton instance of type LightTimeCorrection, used to apply light-time (planetary aberration) corrections when computing vectors from the FrameSystem.

See also

See also vector3 and vector6.

FrameTransformations.Frames.PlanetaryAberration — Constant

PlanetaryAberration

The singleton instance of type PlanetaryAberrationCorrection, used to apply one-way light-time and stellar aberration corrections when computing vectors from the FrameSystem.

See also

See also vector3 and vector6.