Frames's Public Documentation

Frame System

FrameTransformations.Frames.FrameSystem — Type

FrameSystem{O, T, S, E}

A FrameSystem instance manages a collection of user-defined FramePointNode and FrameAxesNode objects, enabling efficient computation of arbitrary transformations between them. It is created by specifying the maximum transformation order O, the outputs datatype T and an AbstractTimeScale instance S. Additionally, an AbstractEphemerisProvider instance E can be provided to compute transformations that require ephemeris data.

The following transformation orders are accepted:

1: position

2: position and velocity

3: position, velocity and acceleration

4: position, velocity, acceleration and jerk

By specifying the maximum transformation the FrameSystem memory usage and performance can be optimised and tailored to the user's needs.

FrameSystem{O, T}()

Create a FrameSystem object of order O and datatype T. The BarycentricDynamicalTime is automatically assigned as the default time scale. The resulting object is constructed with a NullEphemerisProvider, which does not allow the computation of transformation that involve ephemeris files.

Examples

julia> F = FrameSystem{2, Float64}();

julia> @axes ICRF 1 

julia> @axes ECLIPJ2000 17 

julia> add_axes_inertial!(F, ICRF)

julia> add_axes_eclipj2000!(F, ECLIPJ2000, ICRF)

julia> rotation6(F, ICRF, ECLIPJ2000, 0.0)
Rotation{2, Float64}
[...]

julia> rotation9(F, ICRF, ECLIPJ2000, 0.0)
ERROR: Insufficient frame system order: transformation requires at least order 3.

FrameSystem{O, T, S}()

Create a FrameSystem object of order O, datatype T and time scale S. The resulting object is constructed with a NullEphemerisProvider, which does not allow the computation of transformation that involve ephemeris files.

Examples

julia> F = FrameSystem{1, Float64, TerrestrialTime}();

julia> @axes ICRF 1 

julia> @axes ECLIPJ2000 17 

julia> add_axes_inertial!(F, ICRF)

julia> add_axes_eclipj2000!(F, ECLIPJ2000, ICRF)

julia> ep_tt = Epoch("2023-02-10T12:00:00 TT")
2023-02-10T12:00:00.000 TT

julia> rotation3(F, ICRF, ECLIPJ2000, ep_tt)
Rotation{1,Float64}([...])

julia> ep_tdb = Epoch("2023-02-10T12:00:00 TDB")
2023-02-10T12:00:00.000 TDB

julia> rotation3(F, ICRF, ECLIPJ2000, ep_tdb)
ERROR: ArgumentError: Incompatible epoch timescale: expected TerrestrialTime, found BarycentricDynamicalTime.
[...]

FrameSystem{O, T}(eph::AbstractEphemerisProvider)

Create a FrameSystem object of order O and datatype T by providing an instance of an AbstractEphemerisProvider subtype. The timescale is automatically set to the one associated to the ephemeris files loaded in eph. This constructor shall be used when the user desires to compute transformations that involve ephemeris data.

Note

All the kernels that will be used must be loaded within eph. Once the FrameSystem has been created, no additional kernel can be added nor removed.

Examples

julia> using Ephemerides 

julia> eph = EphemerisProvider(DE440_KERNEL_PATH);

julia> F = FrameSystem{2, Float64}(eph)
FrameSystem{2, Float64, BarycentricDynamicalTime, EphemerisProvider}(
  eph: 1-kernel EphemerisProvider,
  points: EMPTY
  axes: EMPTY
)

See also

See also add_axes_inertial!, add_point_root!, vector3 and rotation3

Axes

FrameTransformations.Frames.@axes — Macro

@axes(name, id, type=nothing)

Define a new axes instance to alias the given id. This macro creates an AbstractFrameAxes subtype and its singleton instance callen name. Its type name is obtained by appending Axes to either name or type (if provided).

Examples

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> typeof(ICRF)
InternationalCelestialReferenceFrameAxes

julia> axes_alias(ICRF) 
1

julia> @axes IAU_EARTH 10013

julia> typeof(IAU_EARTH)
IauEarthAxes

See also

See also @point and axes_alias.

FrameTransformations.Frames.axes_alias — Function

axes_alias(ax::AbstractFrameAxes)

Return the axes ID.

FrameTransformations.Frames.is_inertial — Function

is_inertial(frame::FrameSystem, axes::AbstractFrameAxes)
is_inertial(frame::FrameSystem, axesid::Int)

Return true if the given axes are inertial, i.e., non rotating with respect to the root inertial axes.

Note

FixedOffsetAxes with respect to an inertial set of axes, are also consired inertial.

FrameTransformations.Frames.is_timefixed — Function

is_timefixed(frame::FrameSystem, axes::AbstractFrameAxes)
is_timefixed(frame::FrameSystem, axesid::Int)

Return true if the given axes are time-fixed, i.e., their orientation does not change in time with respect to the root inertial axes.

Note

Only :InertialAxes and :FixedOffsetAxes defined with respect to other inertial axes are here considered as time fixed.

FrameTransformations.Frames.ComputableAxesVector — Type

ComputableAxesVector(from, to, order::Int)

Store the properties required to retrieve the i-th order components of a desired vector. Arguments from and to are the NAIFIDs or AbstractFramePoint instances that define the observer and target points.

Only orders between 1 and 3 are supported.

Example

julia> @point SSB 0 SolarSystemBarycenter

julia> @point Sun 10 

julia> ComputableAxesVector(SSB, Sun, 1)
ComputableAxesVector(0, 10, 1)

julia> ComputableAxesVector(0, 10, 1)
ComputableAxesVector(0, 10, 1)

FrameTransformations.Frames.add_axes_bci2000! — Function

add_axes_bci2000!(frames, axes::AbstractFrameAxes, center, data)

Add axes as a set of Body-Centered Inertial (BCI) axes at J2000 relative to the frames system. The center point (i.e., the reference body) is center and can either be the point ID or an AbstractFramePoint instance. data is a dictionary containing a parsed TPC file.

Warning

The parent axes are automatically set to the ICRF (ID = 1). If the ICRF is not defined in frames, an error is thrown.

add_axes_bci2000!(frames, name::Symbol, axesid::Int, cid::Int, data)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also add_axes_fixedoffset!, add_axes_bcrtod! and Orient.AXESID_ICRF.

FrameTransformations.Frames.add_axes_bcrtod! — Function

add_axes_bcrtod!(frames, axes::AbstractFrameAxes, center::AbstractFramePoint, data)

Add axes as a set of Body-Centered Rotating (BCR), True-of-Date (TOD) axes to the frames system. The center point (i.e., the reference body) is center. data is a dictionary containing a parsed TPC file. These axes are the equivalent of SPICE's IAU_<BODY_NAME> frames.

Warning

The parent axes are automatically set to the ICRF (ID = 1). If the ICRF is not defined in frames, an error is thrown.

add_axes_bcrtod!(frames, name::Symbol, axesid::Int, cname::Symbol, cid::Int, data)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro and of an AbstractFramePoint via the @point macro.

See also

See also add_axes_rotating!, add_axes_bci2000! and Orient.AXESID_ICRF.

FrameTransformations.Frames.add_axes_computable! — Function

add_axes_computable!(frames, axes, parent, v1, v2, seq::Symbol)

Add axes as a set of computable axes to frames. Computable axes differ from rotating axes because they are computed through two vectors that are defined within the frame system itself. Computable axes are the equivalent of SPICE's parameterized two-vector frames.

These axes are built such that the first vector, known as the primary vector, is parallel to one axis of the frame; the component of the secondary vector orthogonal to the first is parallel to another axis of the frame, and the cross product of the two vectors is parallel to the remaining axis.

The primary and secondary vectors, v1 and v2 are instances of ComputableAxesVector, which is used to define the NAIF IDs of the vector origin and target, and its order. Current accepted order values are: 1 (position), 2 (velocity) and 3 (acceleration).

For example, to define a primary vector that is parallel to the Sun's (NAIF ID = 10) velocity with respect to the Solary system barycenter (NAIF ID = 0), v1 must be set as: v1 = ComputableAxesVector(10, 0, 1).

seq is a combination of two letters that is used to identify the desired pointing directions of the primary and secondary vectors. Accepted sequences are: :XY, :YX, :XZ, :ZX, :YZ and :ZY.

Given a spacecraft registered as a point in the frame system, an example of a set of computable axes is the Local Vertical Local Horizon (LVLH), where the spacecraf's nadir direction and velocity direction define the axes orientation.

Note

Regardless of the original set of axes in which the primary and secondary vectors are defined, the axes orientation is automatically computed by rotating them to parent.

add_axes_computable!(frames, name::Symbol, axesid::Int, parentid::Int, v1, v2, seq)

Low-level function to add axes name with id axesid to frames as computable axes without requiring the creation of an AbstractFrameAxes type via the @axes macro.

Examples

julia> using Ephemerides 

julia> eph = EphemerisProvider(DE440_KERNEL_PATH);

julia> FRAMES = FrameSystem{4, Float64}(eph);

julia> @point SSB 0 SolarySystemBarycenter 

julia> @point Sun 10 

julia> @axes ICRF 1 InternationalCelestialReferenceFrame 

julia> add_axes_inertial!(FRAMES, ICRF)

julia> add_point_root!(FRAMES, SSB, ICRF)

julia> add_point_ephemeris!(FRAMES, Sun)

julia> @axes SunFrame 2

julia> v1 = ComputableAxesVector(10, 0, 1)
ComputableAxesVector(10, 0, 1)

julia> v2 = ComputableAxesVector(10, 0, 2)
ComputableAxesVector(10, 0, 2)

julia> add_axes_computable!(FRAMES, SunFrame, ICRF, v1, v2, :XY)

See also

See also ComputableAxesVector, add_axes_fixedoffset!, add_axes_inertial! and add_axes_computable!

FrameTransformations.Frames.add_axes_eclipj2000! — Function

add_axes_eclipj2000!(frames, axes::AbstractFrameAxes, parent, iau_model::IAUModel=iau1980)

Add axes as a set of inertial axes representing the Ecliptic Equinox of J2000 (ECLIPJ2000) to frames. The obliquity of the ecliptic is computed using the IAU Model iau_model.

The admissed parent set of axes are the following:

ICRF: for the International Celestial Reference Frame, with ID = 1

MEME2000: the Mean Earth/Moon Ephemeris of J2000, with ID = 22

add_axes_eclipj2000!(frames, name::Symbol, parentid::Int, iau_model::IAUModel=iau1980, 
    axesid::Int = Orient.AXESID_ECLIPJ2000)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

Examples

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> @axes ECLIPJ2000 17 EclipticEquinoxJ2000 

julia> FRAMES = FrameSystem{1, Float64}();

julia> add_axes_inertial!(FRAMES, ICRF)

julia> add_axes_eclipj2000!(FRAMES, ECLIPJ2000, ICRF)

See also

See also add_axes_inertial! and Orient.DCM_ICRF_TO_MEME2000

FrameTransformations.Frames.add_axes_ephemeris! — Function

add_axes_ephemeris!(frames, axes, rot_seq::Symbol)
add_axes_ephemeris!(frames, axes, fun, δfun=nothing, δ²fun=nothing, δ³fun=nothing)

Add axes as a set of ephemeris axes to frames. The orientation of these axes is computed with a series of 3 rotations specified by rot_seq. The euler angles and their derivatives needed by the rotation are extracted from the ephemeris kernels loaded in frames. The parent axes are automatically assigned to the axes with respect to which the orientation data has been written in the kernels.

This operation is only possible if the ephemeris kernels loaded within frames contain orientation data for the AXES ID associated to axes. An error is returned if the parent axes ID is yet to be added to frames.

The rotation sequence is defined by a Symbol specifing the rotation axes. This function assigns dcm = A3 * A2 * A1 in which Ai is the DCM related with the i-th rotation. The possible rot_seq values are: :XYX, XYZ, :XZX, :XZY, :YXY, :YXZ, :YZX, :YZY, :ZXY, :ZXZ, :ZYX, or :ZYZ.

Alternatively, one can provide custom functions that take a vector of euler angles and their derivatives and return the DCM and its derivatives.

The input functions must accept only time as argument and their outputs must be as follows:

fun: return a Direction Cosine Matrix (DCM).

δfun: return the DCM and its 1st order time derivative.

δ²fun: return the DCM and its 1st and 2nd order time derivatives.

δ³fun: return the DCM and its 1st, 2nd and 3rd order time derivatives.

Warning

It is expected that the axes ID assigned by the user are aligned with those used to generate the ephemeris kernels. No check are performed on whether these IDs represent the same physical axes that are intended in the kernels.

add_axes_ephemeris!(frames, name::Symbol, axesid::Int, rot_seq::Symbol)
add_axes_ephemeris!(frames, name::Symbol, axesid::Int, fun, δfun=nothing, δ²fun=nothing, δ³fun=nothing)

Low-level functions to add axes name with id axesid to frames as ephemeris axes without requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also ComputableAxesVector, add_axes_fixedoffset!, add_axes_inertial! and add_axes_computable!

FrameTransformations.Frames.add_axes_fixedoffset! — Function

add_axes_fixedoffset!(frames::FrameSystem, axes, parent, dcm::DCM)

Add axes as a set of fixed offset axes to frames. Fixed offset axes have a constant orientation with respect to their parent axes, represented by dcm, a Direction Cosine Matrix (DCM).

Note

While inertial axes do not rotate with respect to the star background, fixed offset axes are only constant with respect to their parent axes, but might be rotating with respect to some other inertial axes.

add_axes_fixedoffset!(frames, name::Symbol, axesid::Int, parentid::Int, dcm:DCM)

Low-level function to add axes name with id axesid to frames with a fixed-offset from parentid without requiring the creation of an AbstractFrameAxes type via the @axes macro.

Examples

julia> FRAMES = FrameSystem{1, Float64}();

julia> @axes ICRF 1 InternationalCelestialReferenceFrame 

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @axes ECLIPJ2000 17 

julia> add_axes_fixedoffset!(FRAMES, ECLIPJ2000, ICRF, angle_to_dcm(π/3, :Z))

See also

See also add_axes_rotating!, add_axes_inertial! and add_axes_computable!

FrameTransformations.Frames.add_axes_gcrf! — Function

add_axes_gcrf!(frames::FrameSystem)

Add the Geocentric Celestial Reference Frame (GCRF) to the frames graph. The axes are automatically named GCRF and assigned the 23 ID. These axes can only be defined as a set of root axes or as child of the ICRF (ID = 1).

See also

See also add_axes_inertial!, add_axes_icrf! and Orient.AXESID_GCRF.

FrameTransformations.Frames.add_axes_inertial! — Function

add_axes_inertial!(frames, axes; parent=nothing, dcm=nothing)

Add axes as a set of inertial axes to frames. Only inertial axes can be used as root axes to initialise the axes graph. Only after the addition of a set of inertial axes, other axes classes may be added aswell. Once a set of root-axes has been added, parent and dcm become mandatory fields.

Note

The parent of a set of inertial axes must also be inertial. FixedOffset axes that are that only have inertial parents are also accepted.

add_axes_inertial!(frames, name::Symbol, axesid::Int; parentid=nothing, dcm=nothing)

Low-level function to add axes name with id axesid to frames without requiring the creation of an AbstractFrameAxes type via the @axes macro.

Examples

julia> FRAMES = FrameSystem{2, Float64}();

julia> @axes ICRF 1 InternationalCelestialReferenceFrame 

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @axes ECLIPJ2000 17 

julia> add_axes_inertial!(FRAMES, ECLIPJ2000)
ERROR: A set of parent axes for ECLIPJ2000 is required
[...]

julia> add_axes_inertial!(FRAMES, ECLIPJ2000; parent=ICRF, dcm=angle_to_dcm(π/3, :Z))

See also

See also add_axes_rotating!, add_axes_fixedoffset! and add_axes_computable!

FrameTransformations.Frames.add_axes_icrf! — Function

add_axes_icrf!(frames::FrameSystem)

Add the International Celestial Reference Frame (ICRF) as the root axes of the frames graph. The axes are automatically named ICRF and assigned the 1 ID.

See also

See also add_axes_inertial!, add_axes_gcrf! and Orient.AXESID_ICRF.

FrameTransformations.Frames.add_axes_itrf! — Function

add_axes_itrf!(frames, axes::AbstractFrameAxes, parent, model::IAU2006Model=iau2006b)

Add axes as a set of axes representing the International Terrestrial Reference Frame (ITRF) to frames. Use the model argument to specify which IAU model model should be used for the computations. The default is set to iau2006b.

Warning

If the ID of the parent set of axes is neither the ICRF (ID = 1) nor the GCRF (ID = 23), an error is thrown.

add_axes_itrf!(frames, name::Symbol, parentid::Int, model::IAU2006Model=iau2006b, 
    axesid::Int = Orient.AXESID_ITRF)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also add_axes_rotating! and orient_rot3_itrf_to_gcrf.

FrameTransformations.Frames.add_axes_me421! — Function

add_axes_me421!(frames, axes::AbstractFrameAxes, parent)

Add axes as fixed offset axes representing the DE421 Moon's Mean Earth/Mean Rotation (ME) to frames.

Warning

The parent set of axes must either the DE440 Principal Axes (PA440, ID =

or the DE421 Principal Axes (PA421, ID =

31006), otherwise an error is thrown. Depending on that, the relative axes orientation will be automatically selected by this function.

add_axes_me421!(frames, name::Symbol, parentid::Int, axesid::Int = Orient.AXESID_MOONME_DE421)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also add_axes_pa440!, and add_axes_pa421!, Orient.DCM_MOON_PA421_TO_ME421 and Orient.DCM_MOON_PA421_TO_ME421,

FrameTransformations.Frames.add_axes_meme2000! — Function

add_axes_meme2000!(frames, axes::AbstractFrameAxes, parent)

Add axes as a set of inertial axes representing the Mean Equator Mean Equinox of J2000 to frames.

Warning

The the axes ID of the parent set of axes must be 1 (ICRF) or 17 (ECLIPJ2000) otherwise and error is thrown.

add_axes_meme2000!(frames, name::Symbol, parentid::Int, axesid::Int = Orient.AXESID_MEME2000)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

Examples

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> @axes MEME2000 22 MeanEquatorMeanEquinoxJ2000 

julia> FRAMES = FrameSystem{1, Float64}();

julia> add_axes_inertial!(FRAMES, ICRF)

julia> add_axes_meme2000!(FRAMES, MEME2000, ICRF)

See also

See also add_axes_inertial! and Orient.DCM_ICRF_TO_MEME2000

FrameTransformations.Frames.add_axes_mod! — Function

add_axes_mod!(frames, axes::AbstractFrameAxes, parent)

Add axes as a set of projected axes representing the Mean Equator and Equinox of Date (MOD) to frames.

Note

Despite this class of axes has a rotation matrix that depends on time, its derivatives are assumed null.

Warning

The ID of the parent set of axes must be 1 (ICRF), otherwise an error is thrown.

add_axes_mod!(frames, name::Symbol, axesid::Int, parentid::Int)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also add_axes_projected! and Orient.orient_rot3_icrf_to_mod

FrameTransformations.Frames.add_axes_tod! — Function

add_axes_tod!(frames, axes::AbstractFrameAxes, parent)

Add axes as a set of projected axes representing the True Equator of Date (TOD) to frames.

Note

Despite this class of axes has a rotation matrix that depends on time, its derivatives are assumed null.

Warning

The ID of the parent set of axes must be 1 (ICRF), otherwise an error is thrown.

add_axes_tod!(frames, name::Symbol, axesid::Int, parentid::Int)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also add_axes_projected! and Orient.orient_rot3_icrf_to_tod

FrameTransformations.Frames.add_axes_teme! — Function

add_axes_teme!(frames, axes::AbstractFrameAxes, parent)

Add axes as a set of projected axes representing the True Equator, Mean Equinox (TEME) to frames.

Note

Despite this class of axes has a rotation matrix that depends on time, its derivatives are assumed null.

Warning

The ID of the parent set of axes must be 1 (ICRF), otherwise an error is thrown.

add_axes_teme!(frames, name::Symbol, axesid::Int, parentid::Int)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also add_axes_projected! and Orient.orient_rot3_icrf_to_teme

FrameTransformations.Frames.add_axes_pef! — Function

add_axes_pef!(frames, axes::AbstractFrameAxes, parent)

Add axes as a set of rotating axes representing the Pseudo Earth Fixed (PEF) axes to frames.

Warning

The ID of the parent set of axes must be 1 (ICRF), otherwise an error is thrown.

add_axes_pef!(frames, name::Symbol, axesid::Int, parentid::Int)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also add_axes_rotating!, Orient.orient_rot3_icrf_to_pef and Orient.orient_rot6_icrf_to_pef.

FrameTransformations.Frames.add_axes_pa421! — Function

add_axes_pa421!(frames, axes::AbstractFrameAxes)

Add axes as a set of ephemeris axes representing the DE421 Moon's Principal Axes (PA) to frames. The libration angles are extracted from the ephemeris kernels loaded within frames, an error is thrown if such orientation data is not available.

Warning

The parent axes are automatically set to the ICRF (ID = 1). If such axes are not registered in the frame system, an error is thrown.

Warning

To properly read the ephemeris kernels, the ID associated to the input axes must match NAIF's FRAME ID for the Moon PA DE421 axes (31006).

add_axes_pa421!(frames, name::Symbol, axesid::Int = Orient.AXESID_MOONPA_DE421)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also Orient.AXESID_MOONPA_DE421, Orient.orient_rot3_icrf_to_pa421, add_axes_pa440!, and add_axes_me421!

FrameTransformations.Frames.add_axes_pa440! — Function

add_axes_pa440!(frames, axes::AbstractFrameAxes)

Add axes as a set of ephemeris axes representing the DE440 Moon's Principal Axes (PA) to frames. The libration angles are extracted from the ephemeris kernels loaded within frames, an error is thrown if such orientation data is not available.

Warning

The parent axes are automatically set to the ICRF (ID = 1). If such axes are not registered in the frame system, an error is thrown.

Warning

To properly read the ephemeris kernels, the ID associated to the input axes must match NAIF's FRAME ID for the Moon PA DE440 axes (31008).

add_axes_pa440!(frames, name::Symbol, axesid::Int = Orient.AXESID_MOONPA_DE440)

Low-level function to avoid requiring the creation of an AbstractFrameAxes type via the @axes macro.

See also

See also Orient.AXESID_MOONPA_DE440, Orient.orient_rot3_icrf_to_pa440 and add_axes_pa421!.

FrameTransformations.Frames.add_axes_projected! — Function

add_axes_projected!(frames, axes, parent, fun)

Add axes as a set of projected axes to frames. The orientation of these axes depends only on time and is computed through the custom functions provided by the user.

Projected axes are similar to rotating axis, except that all the positions, velocity, etc ... are rotated by the 0-order rotation (i.e. the derivatives of the rotation matrix are null, despite the rotation depends on time).

Warning

It is expected that the input function and their outputs have the correct signature. This function does not perform any checks on the output types.

add_axes_projected!(frames, name::Symbol, axesid::Int, parentid::Int, fun)

Low-level function to add axes name with id axesid to frames as projected axes without requiring the creation of an AbstractFrameAxes type via the @axes macro.

FrameTransformations.Frames.add_axes_rotating! — Function

add_axes_rotating!(frames, axes, parent, fun, δfun=nothing, δ²fun=nothing, δ³fun=nothing)

Add axes as a set of rotating axes to frames. The orientation of these axes depends only on time and is computed through the custom functions provided by the user.

The input functions must accept only time as argument and their outputs must be as follows:

fun: return a Direction Cosine Matrix (DCM).

δfun: return the DCM and its 1st order time derivative.

δ²fun: return the DCM and its 1st and 2nd order time derivatives.

δ³fun: return the DCM and its 1st, 2nd and 3rd order time derivatives.

If δfun, δ²fun or δ³fun are not provided, they are computed via automatic differentiation.

Warning

It is expected that the input functions and their outputs have the correct signature. This function does not perform any checks on the output types.

add_axes_rotating!(frames, name::Symbol, axesid::Int, parentid::Int, fun, δfun=nothing, 
    δ²fun=nothing, δ³fun=nothing)

Low-level function to add axes name with id axesid to frames as rotating axes without requiring the creation of an AbstractFrameAxes type via the @axes macro.

Examples

julia> FRAMES = FrameSystem{3, Float64}();

julia> @axes Inertial 1

julia> add_axes_inertial!(FRAMES, Inertial)

julia> @axes Synodic 2 

julia> fun(t) = angle_to_dcm(t, :Z);

julia> add_axes_rotating!(FRAMES, Synodic, Inertial, fun)

julia> R = rotation6(FRAMES, Inertial, Synodic, π/6);

julia> R[1]
DCM{Float64}:
  0.866025  0.5       0.0
 -0.5       0.866025  0.0
  0.0       0.0       1.0

julia> R[2]
DCM{Float64}:
 -0.5        0.866025  0.0
 -0.866025  -0.5       0.0
  0.0        0.0       0.0

See also

See also add_axes_fixedoffset!, add_axes_inertial! and add_axes_computable!

FrameTransformations.Frames.add_axes_topocentric! — Function

add_axes_topocentric!(frames, axes::AbstractFrameAxes, parent, λ::Number, ϕ::Number, type::Symbol)

Add axes as a set of fixed-offset topocentric axes to frames. The orientation relative to the parent axes parent is defined throuh the longitude λ, the geodetic latitude ϕ and the type type, which may be any of the following:

:NED (North, East, Down): the X-axis points North, the Y-axis is directed eastward and the Z-axis points inwards towards the nadir.

:SEZ (South, East, Zenith): the X-axis points South, the Y-axis is directed East, and the Z-axis points outwards towards the zenith.

:ENU (East, North, Up): the X-axis points East, the Y-axis is directed North and the Z-axis points outwards towards the zenith.

Warning

The parent axes must be a set of body-fixed reference axes. When this constraint is not satisfied, the results may be fundamentally wrong.

add_axes_topocentric!(frames, name::Symbol, axesid::Int, parentid::Int, λ, ϕ, type)

Low-level function to avoid requiring the creation of an AbstractFrameAxes via the @axes macro.

See also

See also add_axes_fixedoffset! and add_point_surface!.

Points

FrameTransformations.Frames.@point — Macro

@point(name, id, type=nothing)

Define a new point instance to alias the given NAIFID id. This macro creates an AbstractFramePoint subtype and its singleton instance called name. Its type name is obtained by appending Point to either name or type (if provided).

Examples

julia> @point Venus 299

julia> typeof(Venus)
VenusPoint 

julia> point_alias(Venus)
299

julia> @point EMB 3 EarthMoonBarycenter

julia> typeof(EMB) 
EarthMoonBarycenterPoint

julia> point_alias(EMB) 
3 

See also

See also @axes and point_alias.

FrameTransformations.Frames.point_alias — Function

point_alias(ax::AbstractFramePoint)

Return the NAIF ID associated to the input point.

FrameTransformations.Frames.add_point_dynamical! — Function

add_point_dynamical!(frames, point::AbstractFramePoint, parent, axes, fun, δfun=nothing, 
    δ²fun=nothing, δ³fun=nothing)

Add point as a time point to frames. The state vector for these points depends only on time and is computed through the custom functions provided by the user. axes is the ID or AbstractFrameAxes instance in which the point state-vector is expressed.

The input functions must accept only time as argument and their outputs must be as follows:

fun: return a 3-elements vector: position

δfun: return a 6-elements vector: position and velocity

δ²fun: return a 9-elements vector: position, velocity and acceleration

δ³fun: return a 12-elements vector: position, velocity, acceleration and jerk

If δfun, δ²fun or δ³fun are not provided, they are computed with automatic differentiation.

Warning

It is expected that the input functions and their ouputs have the correct signature. This function does not perform any checks on whether the returned vectors have the appropriate dimensions.

add_point_dynamical!(frames, name::Symbol, pointid::Int, parentid::Int, axes, fun, 
    δfun=nothing, δ²fun=nothing, δ³fun=nothing)

Low-level function to add point name with ID id to frames as a dynamical point without requiring the creation of an AbstractFramePoint type via the @point macro. axes is the ID or AbstractFrameAxes instance in which the point state-vector is expressed.

Examples

julia> FRAMES = FrameSystem{2, Float64}()

julia> @axes ICRF 1 

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point Origin 0 

julia> add_point_root!(FRAMES, Origin, ICRF)

julia> @point Satellite 1 

julia> satellite_pos(t::T) where T = [cos(t), sin(t), 0]

julia> add_point_dynamical!(FRAMES, Satellite, Origin, ICRF, satellite_pos)

julia> vector6(FRAMES, Origin, Satellite, ICRF, π/6)
6-element SVector{6, Float64} with indices SOneTo(6):
  0.8660254037844387
  0.49999999999999994
  0.0
 -0.49999999999999994
  0.8660254037844387
  0.0

See also

See also add_point_root!, add_point_ephemeris!,add_point_fixed! and add_point_updatable!

FrameTransformations.Frames.add_point_ephemeris! — Function

add_point_ephemeris!(frames::FrameSystem, point::AbstractFramePoint)

Add point as an ephemeris point to frames. This function is intended for points whose state-vector is read from ephemeris kernels (i.e., de440.bsp). The parent point is automatically assigned to the point with respect to which the ephemeris data is written in the kernels. If that point is not available as an ephemeris point in the frame system, an error is thrown.

Note

The axes in which the state-vector is expressed are taken from the ephemeris data: an error is returned if the axes ID is yet to be added to frames.

Warning

It is expected that the NAIF ID and the axes ID assigned by the user are aligned with those used to generate the ephemeris kernels. No check are performed on whether these IDs represent the same physical bodies and axes that are intended in the kernels.

add_point_ephemeris!(frames::FrameSystem, name::Symbol, naifid::Int)

Low-level function to add point name with ID id to frames as an ephemeris point without requiring the creation of an AbstractFramePoint type via the @point macro.

Examples

julia> using Ephemerides 

julia> eph = EphemerisProvider(DE440_KERNEL_PATH);

julia> FRAMES = FrameSystem{2, Float64}(eph);

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point SSB 0 SolarSystemBarycenter

julia> @point Sun 10 

julia> add_point_root!(FRAMES, SSB, ICRF)

julia> add_point_ephemeris!(FRAMES, Sun)

julia> @point Jupiter 599

julia> add_point_ephemeris!(FRAMES, Jupiter)
ERROR: ArgumentError: Ephemeris data for NAIFID 599 is not available in the kernels loaded in the given FrameSystem.
[...]

See also

See also add_point_root!, add_point_fixed!, add_point_dynamical! and add_point_updatable!

FrameTransformations.Frames.add_point_fixed! — Function

add_point_fixed!(frames, point::AbstractFramePoint, parent, axes, offset::AbstractVector)

Add point as a fixed point to frames. Fixed points are those whose positions have a constant offset with respect their parent points in the given set of axes. Thus, points eligible for this class must have null velocity and acceleration with respect to parent. axes is the ID or AbstractFrameAxes instance in which the point state-vector is expressed.

add_point_fixed!(frames, name::Symbol, pointid::Int, parentid::Int, axes, offset::AbstractVector)

Low-level function to add point name with ID id to frames as a fixed-point with respect to point parentid without requiring the creation of an AbstractFramePoint type via the @point macro. axes is the ID or AbstractFrameAxes instance in which the point state-vector is expressed.

Examples

julia> FRAMES = FrameSystem{2, Float64}();

julia> @axes SF -3000 SatelliteFrame

julia> add_axes_inertial!(FRAMES, SF)

julia> @point SC -10000 Spacecraft

julia> @point SolarArrayCenter -10001

julia> add_point_root!(FRAMES, SC, SF)

julia> sa_offset = [0.10, 0.15, 0.30];

julia> add_point_fixed!(FRAMES, SolarArrayCenter, SC, SF, sa_offset)

See also

See also add_point_root!, add_point_ephemeris!, add_point_dynamical! and add_point_updatable!

FrameTransformations.Frames.add_point_root! — Function

add_point_root!(frames::FramesSystem, point::AbstractFramePoint, axes)

Add point as a root point to frames to initialize the points graph. Only after the addition of a root point, other points may be added aswell. This point is intended as the origin, i.e., its position will equal (0., 0., 0.). axes is the ID or AbstractFrameAxes instance in which the point state-vector is expressed.

Note

This operation can be performed only once per FrameSystem object: multiple root points in the same graph are both inadmissible and meaningless.

add_point_root!(frames::FrameSystem, name::Symbol, pointid::Int, axes)

Low-level function to add point name with ID pointid to frames as a root-point without requiring the creation of an AbstractFramePoint type via the @point macro. axes is the ID or AbstractFrameAxes instance in which the point state-vector is expressed.

Examples

julia> FRAMES = FrameSystem{2, Float64}();

julia> @axes ICRF 1 InternationalCelestialReferenceFrame

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point SSB 0 SolarSystemBarycenter 

julia> add_point_root!(FRAMES, SSB, ICRF)

julia> @point Sun 10

julia> add_point_root!(FRAMES, Sun, ICRF)
ERROR: ArgumentError: A root-point is already registed in the given FrameSystem.
[...]

See also

See also add_point_ephemeris!, add_point_fixed!, add_point_dynamical! and add_point_updatable!

FrameTransformations.Frames.add_point_surface! — Function

add_point_surface!(frames, point::AbstractFramePoint, parent, axes, λ::Number, 
    ϕ::Number, R::Number, f::Number=0.0, h::Number=0.0)

Add point to frames as a fixed point on the surface of the parent point body. The relative position is specified by the longitude λ, the geodetic latitude ϕ, the reference radius of the ellipsoid R and its flattening f. The altitude over the reference surface of the ellipsoid h defaults to 0.

Warning

axes must be a set of body-fixed reference axes for the body represented by parent. When this constraint is not satisfied, the results may be fundamentally wrong.

add_point_surface!(frames, point::AbstractFramePoint, parent, axes, pck,  λ, ϕ, h::Number=0.0)

In this case, the ellipsoid parameters are extracted from the input TPC kernel pck using the NAIFId associated to the parent point.

add_point_surface!(frames, name::Symbol, pointid::Int, parentid::Int, axesid::Int, 
    λ::Number, ϕ::Number, R::Number, f::Number=0.0, h::Number=0.0,)

Low-level function to avoid requiring the creation of an AbstractFramePoint via the @point macro.

See also

See also add_point_fixed! and add_axes_topocentric!.

FrameTransformations.Frames.add_point_updatable! — Function

add_point_updatable!(frames, point::AbstractFramePoint, parent, axes)

Add point as an updatable point to frames. Differently from all the other classes, the state vector for updatable points (expressed in the set of input axes) must be manually updated before being used for other computations. axes is the ID or AbstractFrameAxes instance in which the point state-vector is expressed.

Note

This class of points becomes particularly useful if the state vector is not known a-priori, e.g., when it is the output of an optimisation process which exploits the frame system.

add_point_updatable!(frames, name::Symbol, pointid::Int, parentid::Int, axes)

Low-level function to add point name with ID id to frames as an updatable point without requiring the creation of an AbstractFramePoint type via the @point macro. axes is the ID or AbstractFrameAxes instance in which the point state-vector is expressed.

Examples

julia> FRAMES = FrameSystem{2, Float64}();

julia> @axes ICRF 1  

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point Origin 0

julia> @point Satellite 1 

julia> add_point_root!(FRAMES, Origin, ICRF)

julia> add_point_updatable!(FRAMES, Satellite, Origin, ICRF)

julia> y = [10000., 200., 300.]

julia> update_point!(FRAMES, Satellite, y, 0.1)

julia> vector3(FRAMES, Origin, Satellite, ICRF, 0.1)
3-element SVector{3, Float64} with indices SOneTo(3):
 10000.0
   200.0
   300.0

julia> vector3(FRAMES, Origin, Satellite, ICRF, 0.2)
ERROR: UpdatablePoint with NAIFId = 1 has not been updated at time 0.2 for order 1

julia> vector6(FRAMES, Origin, Satellite, ICRF, 0.1)
ERROR: UpdatablePoint with NAIFId = 1 has not been updated at time 0.1 for order 2
[...]

See also

See also update_point!, add_point_root!, add_point_ephemeris!, add_point_dynamical! and add_point_fixed!

FrameTransformations.Frames.update_point! — Function

update_point!(frames, point, stv::AbstractVector, epoch::Epoch)

update_point!(frames::FrameSystem, point, stv::AbstractVector, time)

Update the state vector of point at the input time in frames. The only accepted length for the input vector stv are 3, 6, 9 or 12. The order is automatically inferred from the vector length.

Examples

julia> FRAMES = FrameSystem{2, Float64}();
  
julia> @axes ICRF 1  

julia> add_axes_inertial!(FRAMES, ICRF)

julia> @point Origin 0

julia> @point Satellite 1 

julia> add_point_root!(FRAMES, Origin, ICRF)

julia> add_point_updatable!(FRAMES, Satellite, Origin, ICRF)

julia> y = [10000., 200., 300.];

julia> update_point!(FRAMES, Satellite, y, 0.1)

julia> vector3(FRAMES, Origin, Satellite, ICRF, 0.1)
3-element SVector{3, StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
 10000.0
   200.0
   300.0

julia> vector3(FRAMES, Origin, Satellite, ICRF, 0.2)
ERROR: UpdatablePoint with NAIFId = 1 has not been updated at time 0.2 for order 1 
[...]

julia> vector6(FRAMES, Origin, Satellite, ICRF, 0.1)
ERROR: UpdatablePoint with NAIFId = 1 has not been updated at time 0.1 for order 2 
[...] 

See also

See also add_point_updatable!

Rotations

FrameTransformations.Frames.Rotation — Type

Rotation{S, N}

A container to efficiently compute S-th order rotation matrices of type N between two set of axes. It stores the Direction Cosine Matrix (DCM) and its time derivatives up to the (S-1)-th order. Since this type is immutable, the data must be provided upon construction and cannot be mutated later.

The rotation of state vector between two set of axes is computed with an ad-hoc overload of the product operator. For example, a 3rd order Rotation object R, constructed from the DCM A and its time derivatives δA and δ²A rotates a vector v = [p, v, a] as:

̂v = [A*p, δA*p + A*v, δ²A*p + 2δA*v + A*a]

A Rotation object R call always be converted to a SMatrix or a MMatrix by invoking the proper constructor.

Examples

julia> A = angle_to_dcm(π/3, :Z)
DCM{Float64}:
  0.5       0.866025  0.0
 -0.866025  0.5       0.0
  0.0       0.0       1.0

julia> R = Rotation(A);

julia> SM = SMatrix(R)
3×3 SMatrix{3, 3, Float64, 9} with indices SOneTo(3)×SOneTo(3):
  0.5       0.866025  0.0
 -0.866025  0.5       0.0
  0.0       0.0       1.0

julia> MM = MMatrix(R)
3×3 MMatrix{3, 3, Float64, 9} with indices SOneTo(3)×SOneTo(3):
  0.5       0.866025  0.0
 -0.866025  0.5       0.0
  0.0       0.0       1.0

Rotation(dcms::DCM...)

Create a Rotation object from a Direction Cosine Matrix (DCM) and any of its time derivatives. The rotation order is inferred from the number of inputs, while the rotation type is obtained by promoting the DCMs types.

Examples

julia> A = angle_to_dcm(π/3, :Z); 

julia> δA = DCM(0.0I); 

julia> δ²A = DCM(0.0I); 

julia> R = Rotation(A, δA, δ²A); 

julia> typeof(R) 
Rotation{3, Float64}

julia> R2 = Rotation(A, B, C, DCM(0.0im*I)); 

julia> typeof(R2)
Rotation{4, ComplexF64}

Rotation{S}(dcms::DCM...) where S

Create a Rotation object of order S. If the number of dcms is smaller than S, the remaining slots are filled with null DCMs, otherwise if the number of inputs is greater than S, only the first S DCMs are used.

Warning

Usage of this constructor is not recommended as it may yield unexpected results to unexperienced users.

Rotation{S1}(dcms::NTuple{S2, DCM{N}}) where {S1, S2, N}

Create a Rotation object from a tuple of Direction Cosine Matrix (DCM) and its time derivatives. If S1 < S2, only the first S1 DCMs are considered, otherwise the remaining orders are filled with null DCMs.

Examples

julia> A = angle_to_dcm(π/3, :Z);

julia> B = angle_to_dcm(π/4, π/6, :XY);

julia> R3 = Rotation{3}((A,B));

julia> R3[1] == A
true 

julia> R3[3] 
DCM{Float64}:
 0.0  0.0  0.0
 0.0  0.0  0.0
 0.0  0.0  0.0

Rotation{S}(u::UniformScaling{N}) where {S, N}
Rotation{S, N}(u::UniformScaling) where {S, N}

Create an S-order identity Rotation object of type N with identity position rotation and null time derivatives.

Examples

julia> Rotation{1}(1.0I) 
Rotation{1, Float64}(([1.0 0.0 0.0; 0.0 1.0 0.0; 0.0 0.0 1.0],))

julia> Rotation{1, Int64}(I)
Rotation{1, Int64}(([1 0 0; 0 1 0; 0 0 1],))

Rotation{S1}(rot::Rotation{S2, N}) where {S1, S2, N}
Rotation{S1, N}(R::Rotation{S2}) where {S1, S2, N}

Transform a Rotation object of order S2 to order S1 and type N. The behaviour of these functions depends on the values of S1 and S2:

S1 < S2: Only the first S1 components of rot are considered.

S1 > S2: The missing orders are filled with null DCMs.

Examples

julia> A = angle_to_dcm(π/3, :Z);

julia> B = angle_to_dcm(π/4, π/6, :XY);

julia> R1 = Rotation(A, B);

julia> order(R1)
2

julia> R2 = Rotation{1}(R1);

julia> order(R2)
1

julia> R2[1] == A 
true

julia> R3 = Rotation{3}(R1);

julia> R3[3]
DCM{Float64}:
 0.0  0.0  0.0
 0.0  0.0  0.0
 0.0  0.0  0.0

Rotation(m::DCM{N}, ω::AbstractVector) where N

Create a 2nd order Rotation object of type N to rotate between two set of axes a and b from a Direction Cosine Matrix (DCM) and the angular velocity vector ω of b with respect to a, expressed in b

See also

See also rotation3, rotation6 and rotation9.

FrameTransformations.Frames.order — Function

order(R::Rotation{S}) where S

Return the rotation order S.

Base.inv — Function

inv(rot::Rotation)

Compute the invese of the rotation object rot. The operation is efficiently performed by taking the transpose of each rotation matrix within rot.

FrameTransformations.Frames.rotation3 — Function

rotation3(frame::FrameSystem, from, to, ep::Epoch)

Compute the rotation that transforms a 3-elements state vector from one specified set of axes to another at a given epoch.

Requires a frame system of order ≥ 1.

Inputs

frame – The FrameSystem container object

from – ID or instance of the axes to transform from

to – ID or instance of the axes to transform to

ep – Epoch of the rotation. Its timescale must match that of the frame system.

Output

A Rotation object of order 1.

rotation3(frame::FrameSystem, from, to, t::Number)

Compute the rotation that transforms a 3-elements state vector from one specified set of axes to another at a given time t, expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.rotation6 — Function

rotation6(frame::FrameSystem, from, to, ep::Epoch)

Compute the rotation that transforms a 6-elements state vector from one specified set of axes to another at a given epoch.

Requires a frame system of order ≥ 2.

Inputs

frame – The FrameSystem container object

from – ID or instance of the axes to transform from

to – ID or instance of the axes to transform to

ep – Epoch of the rotation. Its timescale must match that of the frame system.

Output

A Rotation object of order 2.

rotation6(frame::FrameSystem, from, to, t::Number)

Compute the rotation that transforms a 6-elements state vector from one specified set of axes to another at a given time t, expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.rotation9 — Function

rotation9(frame::FrameSystem, from, to, ep::Epoch)

Compute the rotation that transforms a 9-elements state vector from one specified set of axes to another at a given epoch.

Requires a frame system of order ≥ 3.

Inputs

frame – The FrameSystem container object

from – ID or instance of the axes to transform from

to – ID or instance of the axes to transform to

ep – Epoch of the rotation. Its timescale must match that of the frame system.

Output

A Rotation object of order 3.

rotation9(frame::FrameSystem, from, to, t::Number)

Compute the rotation that transforms a 9-elements state vector from one specified set of axes to another at a given time t, expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.rotation12 — Function

rotation12(frame::FrameSystem, from, to, ep::Epoch)

Compute the rotation that transforms a 12-elements state vector from one specified set of axes to another at a given epoch.

Requires a frame system of order ≥ 4.

Inputs

frame – The FrameSystem container object

from – ID or instance of the axes to transform from

to – ID or instance of the axes to transform to

ep – Epoch of the rotation. Its timescale must match that of the frame system.

Output

A Rotation object of order 4.

rotation12(frame::FrameSystem, from, to, t::Number)

Compute the rotation that transforms a 12-elements state vector from one specified set of axes to another at a given time t, expressed in seconds since J2000 if ephemerides are used.

Transformations

FrameTransformations.Frames.LightTime — Constant

LightTime

The singleton instance of type LightTimeCorrection, used to apply light-time (planetary aberration) corrections when computing vectors from the FrameSystem.

See also

See also vector3 and vector6.

FrameTransformations.Frames.PlanetaryAberration — Constant

PlanetaryAberration

The singleton instance of type PlanetaryAberrationCorrection, used to apply one-way light-time and stellar aberration corrections when computing vectors from the FrameSystem.

See also

See also vector3 and vector6.

FrameTransformations.Frames.vector3 — Function

vector3(frame::FrameSystem, from, to, axes, ep::Epoch)

Compute 3-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep.

Requires a frame system of order ≥ 1.

Inputs

frame – The FrameSystem container object

from – ID or instance of the observing point

to – ID or instance of the target point

axes – ID or instance of the output state vector axes

ep – Epoch of the observer. Its timescale must match that of the frame system.

vector3(frame, from, to, axes, t::Number)

Compute 3-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t expressed in seconds since J2000 if ephemerides are used.

vector3(frame, from, to, axes, ep::Epoch, ltcorr, dir; <keyword arguments>)

Compute a light-time corrected 3-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep, using the aberration flag ltcorr, which may be any of the following AbstractLightTimeCorrection:

LightTimeCorrection: it applies the one-way light time (planetary aberration) correction, using a Newtonian formulation.

PlanetaryAberrationCorrection: it applies the one-way light time and stellar aberration corrections using a Newtonian fromulation. It modifies the vector obtained with the LightTimeCorrection option to account for the observer velocity with respect to the Solar System Barycenter.

The integer argument dir is used to specify the correction direction, as follows:

-1: for Reception, in which photons depart from the target's location at the light-time corrected epoch ep-lt and arrive at the observer's location at ep.

+1: for Transmission, in which photons depart from the observer's location at ep and arrive at the target's location at the light-time corrected epoch ep+lt.

Keyword Arguments

iters::Int=1: the number of iterations used to find the solution to the light time equation. For the solar system bodies, a solution is usually found in 3 iterations.

axescenter: ID or instance of the center point for axes. This parameter is used only when axes have an orientation that depends on time. In these cases, the point is used to find the time at which the orientation of the axes should be computed. It defaults to from.

Note

If the PlanetaryAberrationCorrection is applied, the frame system must be at least one order higher than that of the requested transformation.

See also

See also LightTime, PlanetaryAberration and vector6.

References

CSPICE Library

vector3(frame, from, to, axes, t::Number, ltcorr, dir; <keyword arguments>)

Compute a light-time corrected 3-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t, expressed in seconds since J2000, using the aberration flag ltcorr and the direction dir.

FrameTransformations.Frames.vector6 — Function

vector6(frame::FrameSystem, from, to, axes, ep::Epoch)

Compute 6-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep.

Requires a frame system of order ≥ 2.

Inputs

frame – The FrameSystem container object

from – ID or instance of the observing point

to – ID or instance of the target point

axes – ID or instance of the output state vector axes

ep – Epoch of the observer. Its timescale must match that of the frame system.

vector6(frame, from, to, axes, t::Number)

Compute 6-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t expressed in seconds since J2000 if ephemerides are used.

vector6(frame, from, to, axes, ep::Epoch, ltcorr, dir; <keyword arguments>)

Compute a light-time corrected 6-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep, using the aberration flag ltcorr, which may be any of the following AbstractLightTimeCorrection:

LightTimeCorrection: it applies the one-way light time (planetary aberration) correction, using a Newtonian formulation.

PlanetaryAberrationCorrection: it applies the one-way light time and stellar aberration corrections using a Newtonian fromulation. It modifies the vector obtained with the LightTimeCorrection option to account for the observer velocity with respect to the Solar System Barycenter.

The integer argument dir is used to specify the correction direction, as follows:

-1: for Reception, in which photons depart from the target's location at the light-time corrected epoch ep-lt and arrive at the observer's location at ep.

+1: for Transmission, in which photons depart from the observer's location at ep and arrive at the target's location at the light-time corrected epoch ep+lt.

Keyword Arguments

iters::Int=1: the number of iterations used to find the solution to the light time equation. For the solar system bodies, a solution is usually found in 3 iterations.

axescenter: ID or instance of the center point for axes. This parameter is used only when axes have an orientation that depends on time. In these cases, the point is used to find the time at which the orientation of the axes should be computed. It defaults to from.

Note

If the PlanetaryAberrationCorrection is applied, the frame system must be at least one order higher than that of the requested transformation.

See also

See also LightTime, PlanetaryAberration and vector6.

References

CSPICE Library

vector6(frame, from, to, axes, t::Number, ltcorr, dir; <keyword arguments>)

Compute a light-time corrected 6-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t, expressed in seconds since J2000, using the aberration flag ltcorr and the direction dir.

FrameTransformations.Frames.vector9 — Function

vector9(frame::FrameSystem, from, to, axes, ep::Epoch)

Compute 9-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep.

Requires a frame system of order ≥ 3.

Inputs

frame – The FrameSystem container object

from – ID or instance of the observing point

to – ID or instance of the target point

axes – ID or instance of the output state vector axes

ep – Epoch of the observer. Its timescale must match that of the frame system.

vector9(frame, from, to, axes, t::Number)

Compute 9-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t expressed in seconds since J2000 if ephemerides are used.

FrameTransformations.Frames.vector12 — Function

vector12(frame::FrameSystem, from, to, axes, ep::Epoch)

Compute 12-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired epoch ep.

Requires a frame system of order ≥ 4.

Inputs

frame – The FrameSystem container object

from – ID or instance of the observing point

to – ID or instance of the target point

axes – ID or instance of the output state vector axes

ep – Epoch of the observer. Its timescale must match that of the frame system.

vector12(frame, from, to, axes, t::Number)

Compute 12-elements state vector of a target point relative to an observing point, in a given set of axes, at the desired time t expressed in seconds since J2000 if ephemerides are used.