Frames's Low-level API

Axes

FrameTransformations.Frames.AbstractFrameAxes — Type

AbstractFrameAxes

Abstract type for all reference frames axes.

FrameTransformations.Frames.FrameAxesNode — Type

FrameAxesNode{O, T, N} <: AbstractJSMDGraphNode

Define a set of axes.

Fields

name – axes name

class – Symbol representing the class of the axes

id – axes ID (equivalent of NAIFId for axes)

parentid – ID of the parent axes

comp – properties for computable axes

R – rotation matrix for fixed relative axes

f – FrameAxesFunctions container

angles – vector storing the libration angles retrived from ephemerides

FrameTransformations.Frames.axes_name — Function

axes_name(axes::AbstractFrameAxes)

Return the name of axes.

FrameTransformations.Frames.axes_id — Function

axes_id(axes::AbstractFrameAxes)

Return the ID associated to axes.

FrameTransformations.Frames.build_axes — Function

build_axes(frames, name, id, class, funs; parentid, dcm, cax_prop)

Create and add a FrameAxesNode to frames based on the input parameters. Current supported classes are: :InertialAxes, :FixedOffsetAxes, :RotatingAxes, :ProjectedAxes, :EphemerisAxes and :ComputableAxes.

Inputs

frames – Target frame system

name – Axes name, must be unique within frames

id – Axes ID, must be unique within frames

class – Axes class.

funs – FrameAxesFunctions object storing the functions to compute the DCM and, eventually, its time derivatives. It must match the type and order of frames.

Keywords

parentid – Axes ID of the parent axes. Not required only for the root axes.

dcm – DCM with respect to the parent axes. Required only for FixedOffsetAxes.

cax_prop – ComputableAxesProperties, required only by ComputableAxes.

Warning

This is a low-level function and is NOT meant to be directly used. Instead, to add a set of axes to the frame system, see add_axes_inertial!, add_axes_rotating!, add_axes_computable!, add_axes_fixedoffset! and add_axes_projected!.

FrameTransformations.Frames.ComputableAxesProperties — Type

ComputableAxesProperties

Store the properties required to retrieve all the vectors required by a computable set of axes.

Points

FrameTransformations.Frames.AbstractFramePoint — Type

AbstractFramePoint

Abstract type for all reference frames points.

FrameTransformations.Frames.FramePointNode — Type

FramePointNode{O, T, N} <: AbstractJSMDGraphNode

Define a frame system point.

Fields

name – point name

class – Symbol representing the class of the point

axesid – ID of the axes in which the point coordinates are expressed

parentid – NAIF ID of the parent point

NAIFId – NAIF ID of the point

stv – vector storing the point state vectors

epochs – vector storing the epochs associated to stv

nzo – last order at which stv has been computed

f – FramePointFunctions container

FrameTransformations.Frames.point_name — Function

point_name(point::AbstractFramePoint)

Return the name of point.

FrameTransformations.Frames.point_id — Function

point_id(point::AbstractFramePoint)

Return the NAIF ID associated to point.

FrameTransformations.Frames.build_point — Function

build_point(frames, name, NAIFId, class, axesid, funs; parentid, offset)

Create and add a FramePointNode to frames based on the input parameters. Current supported point classes are: :RootPoint, :TimePoint, :EphemerisPoint, :FixedPoint and :UpdatablePoint.

Inputs

frames – Target frame system

name – Point name, must be unique within frames

NAIFId – Point NAIF ID, must be unique within frames

class – Point class.

axesid – ID of the axes in which the state vector of the point is expressed.

funs – FramePointFunctions object storing the functions to update the state vectors of the point. It must match the type and order of frames

Keywords

parentid – NAIF ID of the parent point. Not required only for the root point.

offset – Position offset with respect to a parent point. Required only for FixedPoints.

Warning

This is a low-level function and is NOT meant to be directly used. Instead, to add a point to the frame system, see add_point_ephemeris!, add_point_fixed!, etc...

FrameTransformations.Frames._get_comp_axes_vector3 — Function

_get_comp_axes_vector3(frame, v::ComputableAxesVector, axesid, t)

Compute a 3-elements vector in the desired axes at the given time between two points of the frame system. The returned vector depends on the order in v as follows:

1: position

2: velocity

3: acceleration

FrameTransformations.Frames._get_comp_axes_vector6 — Function

_get_comp_axes_vector6(frame, v::ComputableAxesVector, axesid, t)

Compute a 6-elements vector in the desired axes at the given time between two points of the frame system. The returned vector depends on the order in v as follows:

1: position, velocity

2: velocity, acceleration

3: acceleration, jerk

FrameTransformations.Frames._get_comp_axes_vector9 — Function

_get_comp_axes_vector9(frame, v::ComputableAxesVector, axesid, t)

Compute a 9-elements vector in the desired axes at the given time between two points of the frame system. The returned vector depends on the order in v as follows:

1: position, velocity, acceleration

2: velocity, acceleration, jerk

Warning

This function does not support orders of `v' higher than 2, because it would require the computation of vectors of order 5, which is currently not supported.

FrameTransformations.Frames._get_comp_axes_vector12 — Function

_get_comp_axes_vector12(frame, v::ComputableAxesVector, axesid, t)

Compute a 12-elements vector in the desired axes at the given time between two points of the frame system. The returned vector depends on the order in v as follows:

1: position, velocity, acceleration, jerk

Warning

This function does not support orders of `v' higher than 1, because it would require the computation of vectors of order 5 and 6, which are currently not supported.

Two Vectors

FrameTransformations.Frames.twovectors_to_dcm — Function

twovectors_to_dcm(a, b, seq)

Generate a direction cosine matrix from two time-dependent vectors a and b, following the directions specified in seq.

Inputs

a – The primary vector that will be aligned with the first directions specified in seq.

b – The secondary vector. The component of this vector that is orthogonal to the primary vector is aligned with the second direction specified in the sequence seq.

seq – Accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

Warning

The primary and secondary vectors do not have to be orthogonal. However, a great loss of precision happens when the two vectors are almost aligned. This function does not perform any check on the angular separation of the two vectors. The user should ensure that the primary and secondary vector differ of at least 1 milliradian.

FrameTransformations.Frames.twovectors_to_δdcm — Function

twovectors_to_δdcm(a, b, seq)

Compute the time derivative of a direction cosine matrix generated from two time-dependent state vectors a and b, following the directions specified in seq.

Inputs

a and b – 6-elements state vectors (position and velocity).

seq – Accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

FrameTransformations.Frames.twovectors_to_δ³dcm — Function

twovectors_to_δ³dcm(a, b, seq)

Compute the 3rd-order time derivative of a direction cosine matrix generated from two time-dependent state vectors a and b, following the directions specified in seq.

Inputs

a and b – 12-elements state vectors (position, velocity, acceleration and jerk).

seq – Accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

FrameTransformations.Frames.twovectors_to_δ²dcm — Function

twovectors_to_δ²dcm(a, b, seq)

Compute the 2nd-order time derivative of a direction cosine matrix generated from two time-dependent state vectors a and b, following the directions specified in seq.

Inputs

a and b – 9-elements state vectors (position velocity and acceleration).

seq – Accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

FrameTransformations.Frames._twovectors_to_dcm — Function

_twovectors_to_dcm(a, b, seq::Symbol, fc::Function, fn::Function, fk::Function)

Generate a direction cosine matrix and/or its time derivatives from the vectors a and b, according to the directions specifeid in seq.

Notes

fc and fn are used to control the derivative order.

FrameTransformations.Frames._two_vectors_to_rot6 — Function

_two_vectors_to_rot6(a, b, seq::Symbol)

Generate a direction cosine matrix and time-derivative, minimising the number of repeated computations.

See also

See twovectors_to_dcm and twovectors_to_δdcm for more information.

FrameTransformations.Frames._two_vectors_to_rot9 — Function

_two_vectors_to_rot9(a, b, seq::Symbol)

Generate a direction cosine matrix and its 1st and 2nd-order time-derivatives, minimising the number of repeated computations.

See also

See twovectors_to_dcm and twovectors_to_δ²dcm for more information.

FrameTransformations.Frames._two_vectors_to_rot12 — Function

_two_vectors_to_rot12(a, b, seq::Symbol)

Generate a direction cosine matrix and its 1st, 2nd and 3r-order time-derivatives, minimising the number of repeated computations.

See also

See twovectors_to_dcm and twovectors_to_δdcm for more information.

FrameTransformations.Frames._two_vectors_basis — Function

_two_vectors_basis(a, b, seq::Symbol, fc::Function, fk::Function)

Generate a 3D right-handed orthogonal vector basis and/or its time derivatives from the vectors a and b, according to the directions specified in seq and the input cross function fc. fk is a function that filters a to guarantee type-stability.

The accepted sequence directions are: :XY, :YX, :XZ, :ZX, :YZ, :ZY

The standard basis, its 1st, 2nd-order and 3rd order time derivatives can be computed by passing cross, cross6, cross9 or cross12 to fc. The returned vectors will have a length of 3, 6 or 9, respectively.