Orient

API

Types

FrameTransformations.Orient.EOPData — Type

EOPData{T}

EOP Data for IAU 2000A.

Note

Each field will be an AbstractInterpolation indexed by the Julian Day.

Fields

x, y: Polar motion with respect to the crust [arcsec].

UT1_UTC: Irregularities of the rotation angle [s].

LOD: Length of day offset [ms].

dX, dY: Celestial pole offsets referred to the model IAU2000A [milliarcsec].

FrameTransformations.Orient.FundamentalArguments — Type

FundamentalArguments(t::Number, m::IAUModel=iau2006a)

Compute the fundamental luni-solar and planetary arguments, in radians, associated to the given IAU model m, at time t expressed in TDB Julian centuries since J2000.

The available IAU model options are: iau2006a and iau2000b.

Note

Only the planetary arguments are affected by the IAU Model choice.

FrameTransformations.Orient.FundamentalArguments — Type

FundamentalArguments{N <: Number}

Type storing the fundamental luni-solar and planetary arguments.

Fields

Mₐ – Mean anomaly of the Moon

Sₐ – Mean anomaly of the Sun

uₘ – Mean longitude of the Moon minus mean longitude of the ascending node F

Dₛ – Mean elongation of the Moon from the Sun

Ωₘ – Mean longitude of the Moon's ascending node

λ_Me – Mercury's mean heliocentric longitude

λ_Ve – Venus's mean heliocentric longitude

λ_Ea – Earth's mean heliocentric longitude

λ_Ma – Mars's mean heliocentric longitude

λ_Ju – Jupiter's mean heliocentric longitude

λ_Sa – Saturn's mean heliocentric longitude

λ_Ur – Uranus's mean heliocentric longitude

λ_Ne – Neptune's mean heliocentric longitude

pₐ – General precession in longitude

Functions

FrameTransformations.Orient.LuniSolarArguments — Method

LuniSolarArguments(t::Number, m::IAUModel)

Compute the fundamental (Delaunay) Luni-Solar arguments, in radians, associated to the desired IAU model m, at time t expressed in TDB Julian centuries since J2000.

The returned values depend on the input model as follows:

IAU2006A: the Delaunay expressions are taken from the IERS 2010 Conventions.

IAU2000B: the expressions are taken from Simon et al. (1994), following ERFA's implementation of nut00b.c

Outputs

Mₐ – Mean anomaly of the Moon

Sₐ – Mean anomaly of the Sun

uₘ – Mean longitude of the Moon minus mean longitude of the ascending node F

Dₛ – Mean elongation of the Moon from the Sun

Ωₘ – Mean longitude of the Moon's ascending node

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

Simon, J. et al., (1994), Numerical expressions for precession formulae and mean elements for the Moon and the planets.

ERFA software library

FrameTransformations.Orient.PlanetaryArguments — Method

PlanetaryArguments(t::Number)

Compute the fundamental planetary arguments and the general precession, in radians, at time t expressed in TDB Julian centuries since J2000.

Outputs

λ☿ – Mercury's mean heliocentric longitude.

λ♀ – Venus's mean heliocentric longitude.

λe – Earth's mean heliocentric longitude.

λ♂ – Mars's mean heliocentric longitude.

λ♃ – Jupiter's mean heliocentric longitude.

λ♄ – Saturn's mean heliocentric longitude.

λ⛢ – Uranus's mean heliocentric longitude.

λ♆ – Neptune's mean heliocentric longitude.

pλ – General precession in longitude.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

FrameTransformations.Orient.bpn2xy — Method

bpn2xy(A::AbstractMatrix)

Compute the CIP X and Y coordinates from the bias-precession-nutation matrix, in radians.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

FrameTransformations.Orient.cio_locator — Method

cio_locator(m::IAUModel, t::Number, x::Number, y::Number)

Compute the CIO Locator s in radians, according to the IAU Model m, given the CIP coordinates X and Y at time t expressed in TT Julian centuries since J2000

The function has been implemented for the IAU2000, IAU2006 and the CPN models.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

ERFA library

ERFA library

FrameTransformations.Orient.cip_coords — Method

cip_coords(m::IAUModel, t::Number)

Computes the CIP X, Y coordinates, in radians, according to the IAU model m at time t expressed in TT Julian Centuries since J2000.

This function has been implemented for the IAU2000, IAU2006 and the CPN models.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

FrameTransformations.Orient.cip_motion — Function

cip_motion(m::IAUModel, t::Number, dx::Number=0.0, dy::Number=0.0)

Compute the CIRS to GCRS rotation matrix, according to the IAU Model m, at time t expressed in TT Julian centuries since J2000. Optional IERS corrections for free-core nutation and time depedented effects can be provided through dx and dy.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

FrameTransformations.Orient.earth_rotation_angle — Method

earth_rotation_angle(t::Number)

Compute the Earth Rotation Angle (ERA) in radians, i.e., the angle between the Celestial Intermediate Origin (CIO) and the Terrestrial Intermediate Origin (TIO) at time t expressed as UT1 days since J2000.

Note

The function uses the fractional UT1 date to gain additional precision in the computations (0.002737.. instead of 1.002737..)

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

ERFA library

FrameTransformations.Orient.earth_rotation_rate — Method

earth_rotation_rate(LOD::Number)

Compute the true angular velocity of the Earth accounting for the Length of the Day, i.e., the instantaneous rate of change of UT1 with respect to a uniform time scale.

FrameTransformations.Orient.earth_rotation_rate — Method

earth_rotation_rate()

Compute the nominal Earth angular velocity.

FrameTransformations.Orient.era_rotm — Method

era_rotm(t::Number)

Compute the TIRS to CIRS Earth Rotation matrix, according to the IERS 2010 conventions at time t expressed as UT1 days since J2000.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

FrameTransformations.Orient.fa_mano_moon — Method

fa_mano_moon(t::Number)

Return the mean anomaly of the Moon Mₐ, in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mano_sun — Method

fa_mano_sun(t::Number)

Return the mean anomaly of the Sun Sₐ in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_melo_moon — Method

fa_melo_moon(t::Number)

Return the mean elongation of the Moon from the Sun D in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlat_moon — Method

fa_mlat_moon(t::Number)

Return the mean longitude of the Moon minus the mean longitude of the ascending node F in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_earth — Method

fa_mlon_earth(t::Number)

Return the mean heliocentric longitude of Earth in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_jupiter — Method

fa_mlon_jupiter(t::Number)

Return the mean heliocentric longitude of Jupiter in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_mars — Method

fa_mlon_mars(t::Number)

Return the mean heliocentric longitude of Mars in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_mercury — Method

fa_mlon_mercury(t::Number)

Return the mean heliocentric longitude of Mercury in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_moon — Method

fa_mlon_moon(t::Number)

Return the mean longitude of the Moon's ascending node Ω in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_neptune — Method

fa_mlon_neptune(t::Number)

Return the mean heliocentric longitude of Neptune in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_saturn — Method

fa_mlon_saturn(t::Number)

Return the mean heliocentric longitude of Saturn in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_uranus — Method

fa_mlon_uranus(t::Number)

Return the mean heliocentric longitude of Uranus in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_venus — Method

fa_mlon_venus(t::Number)

Return the mean heliocentric longitude of Venus in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_precession — Method

fa_precession(t::Number)

Return the general accumulated precession in longitude pₐ in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.frame_bias — Method

frame_bias(::IAU2000Model)

Compute the frame bias components of the IAU 2000 precession-nutation models, in radians.

Notes

The frame bias corrections in longitude and obliquity are required to correct for the offset between the GCRS pole and the mean J2000 pole. They define, with respect to the GCRS axes, a J2000 mean pole that is consistent with teh IAU 2000A precession-nutation model.

The function also returns an offset in right ascension taken from Chapront et al. (2002), necessary to completely describe the frame bias, but that is not part of the original IAU 2000A model.

References

ERFA software library

FrameTransformations.Orient.fw2xy — Method

fw2xy(ϵ::Number, ψ::Number, γ::Number, φ::Number)

Compute the CIP X and Y coordinates from Fukushima-Williams bias-precession-nutation angles, in radians.

Inputs

ϵ – F-W angle with IAU 2006A/B nutation corrections.

ψ – F-W angle with IAU 2006A/B nutation corrections.

γ – F-W angle

ϕ – F-W angle

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

FrameTransformations.Orient.fw_angles — Method

fw_angles(m::IAU2006Model, t::Number)

Compute the precession angles in radians, following the IAU 2006 Fukushima-Williams 4-angle formulation at time t expressed in TT Julian centuries since J2000.

Outputs

γ – F-W 1st angle

ϕ – F-W 2nd angle

ψ – F-W 3rd angle

ε – F-W 4th angle

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library

FrameTransformations.Orient.fw_matrix — Method

fw_matrix(γ, ϕ, ψ, ε)

Form the Nutation-Precession-Bias (NPB) rotation matrix given the Fukushima-Williams angles, expressed in radians.

The present function can construct three different matrices depending on which angles are supplied as arguments:

NPB: To obtain the Nutation-Precession-Bias matrix, generate the four standard FW precession angles (̄γ, ̄ϕ, ̄ψ, ϵₐ) then generate the nutation components Δψ and Δϵ and add them to ̄ψ, ϵₐ. Finally, call the present functions using those four angles as arguments.

PB: To obtain the precession-frame bias matrix, generate the four standard FW precession angles and call the present function.

B: To obtain the frame bias matrix, generate the four standard FW precession angles at date J2000.0 and call this function.

The remaining nutation-only and precession only matrices can be obtained by combining these three appropriately.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library

FrameTransformations.Orient.get_iers_eop — Method

get_iers_eop(; force_download = false)

Download and parse the IERS EOP C04 data.

The files are downloaded using the RemoteFile package with weekly updates. Hence, if one desires to force a download before the scheduled time, then set the keyword force_download to true.

Note

The files will be downloaded from the default URL. If the user want to use another one, then use the specialized function get_iers_eop_IAU2000A

See also: get_iers_eop_IAU2000A

Returns

A structure EOPData with the interpolations of the EOP parameters. Notice that the interpolation indexing is set to Julian days since J2000.

FrameTransformations.Orient.get_iers_eop_IAU2000A — Function

get_iers_eop_iau_2000A(url::String="https://datacenter.iers.org/data/csv/finals2000A.all.csv"; 
    force_download = false)

Get the IERS EOP C04 IAU2000A data from the URL url.

If url is omitted, then it defaults to https://datacenter.iers.org/data/csv/finals2000A.all.csv. The file is downloaded using the RemoteFile package with weekly updates. Hence, if one desires to force a download before the scheduled time, then set the keyword force_download to true.

Note

The interpolation of every field in EOPData between two points in the grid is linear. If extrapolation is needed, then if will use the nearest value (flat extrapolation).

See also: get_iers_eop

Returns

The structure EOPData with the interpolations of the EOP parameters. Notice that the interpolation indexing is set to Julian days since J2000.

FrameTransformations.Orient.offset_utc2ut1 — Method

offset_utc2ut1(seconds)

Return the offset between UTC and UT1 in seconds.

FrameTransformations.Orient.orient_bias_precession — Method

orient_bias_precession(m::IAUModel, t::Number)

Form the precession-frame bias (PB) matrix that transforms vectors from the GCRS to the mean of date, following the IAU Model m at time t expressed as TT Julian centuries since J2000.

The function has been implemented for the IAU2000 and IAU2006 models.

References:

IAU: Trans. International Astronomical Union, Vol. XXIVB; Proc. 24th General Assembly, Manchester, UK. Resolutions B1.3, B1.6. (2000)

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA pmat06 function.

ERFA pmat00 function.

FrameTransformations.Orient.orient_bias_precession_nutation — Method

orient_bias_precession_nutation(m::IAUModel, t::Number)

Compute the equinox-based bias-precession-nutation matrix using the IAU Model m procedures at time t expressed in TT Julian centuries since J2000.

The function has been implemented for the IAU2000 and IAU2006 models.

Note

The computed matrix rotates a vector from the GCRS to the true equatorial triad of date.

References

ERFA pn06 function

ERFA pn00 function

FrameTransformations.Orient.orient_nutation — Method

orient_nutation(m::IAUModel, t::Number)

Compute the nutation components in longitude and obliquity for the IAU Model m, in radians, at time t expressed in TT Julian Centuries since J2000.

Notes

Due to their theoretical basis, the original developments required t expressed as TDB. However, in practice, it is usually more convenient to use Terrestrial Time (TT) as it makes no significant differences (< 0.01 μas) in the final result.

For the IAU 2006A model, the function strictly follows the SOFA implementation. It first computes the IAU 2000A nutation, then applies adjustments for the consequences of the change in obliquity from the IAU 1980 ecliptic to the IAU 2006 ecliptic and (ii) for the secular variation in the Earth's dynamical form factor J2. These corrections ensure that the IAU 2000A nutation is consistent with the IAU 2006 precession model. Please note that the coefficients available on the IERS tables already include those corrections, and are retrieved by multiplying the amplitudes of the SOFA nutation in longitude coefficients by 1.00000047.

The computation of the free-core nutation and time dependent effects are excluded from this model. To achieve the < 1μas accuracy with the IAU 2006/2000 A precession-nutation models, such effects must be included a-posteriori (through dX and dY) using the IERS EOP data.

For the IAU 2000B model, the nutation series is truncated from nearly 1400 terms to only 77, yet it still delivers results of 1 mas accuracy at present epochs. In particular, it delivers a pole accurate to 1 mas from 1900 to 2100 (only very occasionally just outside 1 mas). The coefficients are taken from SOFA's implementation, which slighlty differ from those reported in McCarthy and Luzum (2003). Comparisons with IAU 2006A show that the SOFA version between 1995 and 2050 delivers 0.283 mas RMSE (0.994 mas in the worst case), whereas the IERS Conventions website version delivers 0.312 RMSE (1.125 mas in the worst case).

The IAU 2000B model includes constant planetary bias terms that compensate for long-period nutations. These amplitudes used in this implementation are optimised for a rigorous method, where frame bias, precession and nutation are applied separately and in that order (see SOFA's documentation for further insights).

A simplified version of the Fundamental Arguments, taken from Simon et al (1994) is exploited for IAU2000B as the error introduced is below the model accuracy ( ~0.1 mas).

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Simon, J. et al., (1994), Numerical expressions for precession formulae and mean elements for the Moon and the planets.

ERFA nut06a and nut00b functions

FrameTransformations.Orient.orient_obliquity — Method

orient_obliquity(m::IAUModel, t::Number)

Compute the mean obliquity of the ecliptic at epoch, in radians, at time t expressed in TT Julian centuries since J2000.

Note

This function is implemented only for IAU1980 and IAU2006 models. IAU 2000 Models implement proper precession-rate corrections to the IAU1980 mean obliquity.

References

ERFA obl80 and obl06 functions.

FrameTransformations.Orient.orient_rot12_itrf_to_gcrf — Function

orient_rot12_itrf_to_gcrf(m::IAUModel, tt, ut1, xₚ, yₚ, dX=0.0, dY=0.0, LOD=0.0)

Compute the rotation matrix from ITRF to GCRF and its derivatives up to order 3, according to the IAU Model m, at time tt and ut1 expressed in TT seconds and UT1 days since J2000, respectively.

This function has been implemented for IAU2000, IAU2006 and CPN models.

Note

All the input quantities xₚ, yₚ, dX and dY must be expressed in radians

FrameTransformations.Orient.orient_rot12_itrf_to_gcrf — Function

orient_rot12_itrf_to_gcrf(m::IAUModel, t::Number)

Compute the rotation matrix from ITRF to GCRF and its time derivatives up to order 3 at time t expressed as TT seconds since J2000, according to the IAU Model m.

FrameTransformations.Orient.orient_rot3_icrf_to_mememod — Method

orient_rot3_icrf_to_mememod(t::Number)

Compute the rotation matrix from the International Celestial Reference Frame (ICRF) to the Mean Equinox Mean Equator of Date at time t, expressed in TT seconds since J2000.

FrameTransformations.Orient.orient_rot3_icrf_to_pa421 — Method

orient_rot3_icrf_to_pa421(eph::AbstractEphemerisProvider, ep::Epoch)

Compute the rotation matrix from the ICRF to the DE421 Moon's Principal Axes at the input epoch ep.

FrameTransformations.Orient.orient_rot3_icrf_to_pa421 — Method

orient_rot3_icrf_to_pa421(eph::AbstractEphemerisProvider, t::Number)

Compute the rotation matrix from the ICRF to the DE421 Moon's Principal Axes at the given input time t, expressed in seconds since J2000.

Warning

This function is not optimised for performance (it allocates!). The user is suggested to retrieve the Principal axes orientation using the dedicated Frame System functions.

FrameTransformations.Orient.orient_rot3_icrf_to_pa440 — Method

orient_rot3_icrf_to_pa440(eph::AbstractEphemerisProvider, ep::Epoch)

Compute the rotation matrix from the ICRF to the DE440 Moon's Principal Axes at the input epoch ep.

FrameTransformations.Orient.orient_rot3_icrf_to_pa440 — Method

orient_rot3_icrf_to_pa440(eph::AbstractEphemerisProvider, t::Number)

Compute the rotation matrix from the ICRF to the DE440 Moon's Principal Axes at the given input time t, expressed in seconds since J2000.

Warning

This function is not optimised for performance (it allocates!). The user is suggested to retrieve the Principal axes orientation using the dedicated Frame System functions.

FrameTransformations.Orient.orient_rot3_itrf_to_gcrf — Function

orient_rot3_itrf_to_gcrf(m::IAUModel, tt, ut1, xₚ, yₚ, dX=0.0, dY=0.0)

Compute the rotation matrix from ITRF to GCRF according to the IAU Model m, at time tt and ut1 expressed in TT seconds and UT1 days since J2000, respectively.

This function has been implemented for IAU2000, IAU2006 and CPN models.

Note

All the input quantities xₚ, yₚ, dX and dY must be expressed in radians

FrameTransformations.Orient.orient_rot3_itrf_to_gcrf — Method

orient_rot3_itrf_to_gcrf(m::IAUModel, t::Number)

Compute the rotation matrix from ITRF to GCRF at time t expressed as TT seconds since J2000, according to the IAU Model m, as follows:

IAU2000A: the pole coordinates (xₚ, yₚ) and the free-core nutation and time corrections to the CIP coordinates (dX, dY) are interpolated from the latest released IERS EOP data. The precession-nutation matrix is computed using the full IAU 2000A model.

IAU2000B: only the pole coordinates (xₚ, yₚ) are interpolated from the latest EOP data. The Free Core Nutation (FCN) corrections dX, dY are neglected. The precession-nutation matrix is computed following the IAU 2000 model but with truncated expressions for the nutation corrections.

IAU2006A: the pole coordinates (xₚ, yₚ) and the free-core nutation and time corrections to the CIP coordinates (dX, dY) are interpolated from the latest released IERS EOP data. The precession-nutation matrix is computed using the full IAU 2006/2000A model.

IAU2006B: only the pole coordinates (xₚ, yₚ) are interpolated from the latest EOP data. The Free Core Nutation (FCN) corrections dX, dY are neglected. The precession-nutation matrix is computed following the IAU 2006A model but with truncated expressions for the nutation corrections.

CPNc: a concise model with a cut-off at 2.5 mas of the X and Y series, delivering a worst-case accuracy of about 15 mas between 1995-2050. It does not take into account the Free Core Nutation (~0.2 mas).

CPNd: an extremely concise formulation with an accuracy of about 1 arcsec between 1995 and 2050. It neglects polar-motion (~0.25 arcsec), the FCN corrections and the CIO locator.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

FrameTransformations.Orient.orient_rot6_itrf_to_gcrf — Function

orient_rot6_itrf_to_gcrf(m::IAUModel, tt, ut1, xₚ, yₚ, dX=0.0, dY=0.0, LOD=0.0)

Compute the rotation matrix from ITRF to GCRF and its derivative, according to the IAU Model m, at time tt and ut1 expressed in TT seconds and UT1 days since J2000, respectively.

This function has been implemented for IAU2000, IAU2006 and CPN models.

Note

All the input quantities xₚ, yₚ, dX and dY must be expressed in radians

FrameTransformations.Orient.orient_rot6_itrf_to_gcrf — Function

orient_rot6_itrf_to_gcrf(m::IAUModel, t::Number)

Compute the rotation matrix from ITRF to GCRF and its derivative at time t expressed as TT seconds since J2000, according to the the IAU Model m.

FrameTransformations.Orient.orient_rot9_itrf_to_gcrf — Function

orient_rot9_itrf_to_gcrf(m::IAUModel, tt, ut1, xₚ, yₚ, dX=0.0, dY=0.0, LOD=0.0)

Compute the rotation matrix from ITRF to GCRF and its derivatives up to order 2, according to the IAU Model m, at time tt and ut1 expressed in TT seconds and UT1 days since J2000, respectively.

This function has been implemented for IAU2000, IAU2006 and CPN models.

Note

All the input quantities xₚ, yₚ, dX and dY must be expressed in radians

FrameTransformations.Orient.orient_rot9_itrf_to_gcrf — Function

orient_rot9_itrf_to_gcrf(m::IAUModel, t::Number)

Compute the rotation matrix from ITRF to GCRF and its time derivatives up to order 2 at time t expressed as TT seconds since J2000, according to the IAU Model m.

FrameTransformations.Orient.polar_motion — Method

polar_motion(t::Number, xₚ::Number, yₚ::Number)

Compute the Polar Motion rotation matrix from ITRF to TIRS, according to the IERS 2010 Conventions, at time t expressed in TT Julian centuries since J2000. The function requires xp and yp, the Celestial Intermediate Pole (CIP) coordinates with respect to the International Celestial Reference Frame (ITFR).

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

FrameTransformations.Orient.precession_angles — Method

precession_angles(m::IAU1980Model, t::Number)

Compute the precession angles from Lieske et al. 1977 model, in radians, at time t expressed in TT Julian centuries since J2000.

References

ERFA software library

FrameTransformations.Orient.precession_rate — Method

precession_rate(m::IAU2000Model, t::Number)

Compute the precession-rate part of the IAU 2000 precession-nutation models, in radians, at time t expressed as TT Julian centuries since J2000.

References

ERFA software library

FrameTransformations.Orient.tio_locator — Method

tio_locator(t::Number)

Compute the TIO locator s' at date, positioning the Terrestrial Intermediate Origin on the equator of the Celestial Intermediate Pole (CIP) at time t expressed as TT Julian centuries since J2000.

This function approximates the unpredictable motion of the TIO locator s' with its secular drift of ~0.47 μas/century.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Lambert, S. and Bizouard C. (2002), Positioning the Terrestrial Ephemeris Origin in the Terrestrial Reference Frame, DOI: 10.1051/0004-6361:20021139

FrameTransformations.Orient.xys2m — Method

xys2m(x::Number, y::Number, s::Number)

Compute the Intermediate-to-Celestial matrix given the CIP x, y' coordinates and the CIO locators`, all in radians.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library

Macros

Constants

FrameTransformations.Orient.AXESID_ECLIPJ2000 — Constant

AXESID_ECLIPJ2000

NAIF Axes ID for the Mean Ecliptic Equinox of J2000 (ECLIPJ2000)

FrameTransformations.Orient.AXESID_ICRF — Constant

AXESID_ICRF

NAIF Axes ID for the International Celestial Reference Frame (ICRF)

FrameTransformations.Orient.AXESID_MEME2000 — Constant

AXESID_MEME2000

Axes ID for the Mean Dynamical Equator and Equinox of J2000.0.

Note

In SPICE the J2000 and ICRF axes are considered equal, thus there exist no specific NAIF ID for the MEME2000 axes. 22 has been chosen because it is the first unassigned axes ID among the built-in SPICE frames.

FrameTransformations.Orient.AXESID_MOONME_DE421 — Constant

AXESID_MOONME_DE421

NAIF axes id for the DE421 Moon Mean Earth/Mean Rotation axes (ME421).

FrameTransformations.Orient.AXESID_MOONPA_DE421 — Constant

AXESID_MOONPA_DE421

NAIF axes id for the DE421 Moon Principal Axes (PA421).

FrameTransformations.Orient.AXESID_MOONPA_DE440 — Constant

AXESID_MOONPA_DE440

NAIF Axes id for the DE440 Moon Principal Axes (PA440).

FrameTransformations.Orient.CPNc — Constant

CPNc

The singleton instance of type CPNC, representing the concise CPNc from Capitaine & Wallace, Concise CIO based precession-nutation formulations, (2008). This model truncates the X, Y series to deliver an accuracy of few mas.

FrameTransformations.Orient.CPNd — Constant

CPNd

The singleton instance of type CPND, representing the concise CPNd from Capitaine & Wallace, Concise CIO based precession-nutation formulations, (2008). This model truncates the X, Y series to deliver an accuracy of few arcseconds.

FrameTransformations.Orient.DCM_ICRF_TO_ECLIPJ2000 — Constant

DCM_ICRF_TO_ECLIPJ2000

DCM for the rotation from the International Celestial Reference Frame (ICRF) to the Mean Ecliptic Equinox of J2000 (ECLIPJ2000).

FrameTransformations.Orient.DCM_ICRF_TO_J2000_BIAS — Constant

DCM_ICRF_TO_J2000_BIAS

DCM for the rotation from the International Celestial Reference Frame (ICRF) and the Mean Dynamical Equator and Equinox of J2000.0 (MEME2000).

Note

The frame bias is here computed using the IAU 2006 Precession model, similarly to ESA's GODOT. Some other software libraries, such as Orekit, use the frame bias of the IAU 2000 precession model. The two definitions differ of about 1 arcsecond.

References

Hilton, James L., and Catherine Y. Hohenkerk. – Rotation matrix from the mean dynamical equator and equinox at J2000. 0 to the ICRS. – Astronomy & Astrophysics 513.2 (2004): 765-770. DOI: 10.1051/0004-6361:20031552

SOFA docs

FrameTransformations.Orient.DCM_J2000_TO_ECLIPJ2000 — Constant

DCM_J2000_TO_ECLIPJ2000

DCM for the rotation from the Mean Dynamical Equator of J2000 (MEME2000) to the Mean Ecliptic Equinox. This corresponds to the transformation J2000 -> ECLIPJ2000 in the SPICE toolkit, and uses the mean obliquity of the ecliptic from the IAU 1976 theory.

FrameTransformations.Orient.DCM_MOON_PA421_TO_ME421 — Constant

DCM_MOON_PA421_TO_ME421

DCM for the rotation from the Moon Principal Axis 421 (PA421) to the Moon Mean Earth/Mean Rotation DE421 (ME421) axes.

References

J. G. Williams et al. (2008), DE421 Lunar Orbit, Physical Librations, and Surface Coordinates, DE421 Lunar Ephemeris and Orientation

FrameTransformations.Orient.DCM_MOON_PA430_TO_ME421 — Constant

DCM_MOON_PA430_TO_ME421

DCM for the rotation from the Moon Principal Axis 430 (PA430) to the Moon Mean Earth/Mean Rotation DE421 (ME421) axes.

References

Folkner M. William et al. (2014), The Planetary and Lunar EphemeridesDE430 and DE431

J. G. Williams et al. (2013), DE430 Lunar Orbit, Physical Librations, and Surface Coordinates, DE430 Lunar Ephemeris and Orientation

FrameTransformations.Orient.DCM_MOON_PA430_TO_ME430 — Constant

DCM_MOON_PA430_TO_ME430

DCM for the rotation from the Moon Principal Axis 430 (PA430) to the Moon Mean Earth/Mean Rotation DE430 (ME430) axes.

References

Folkner M. William et al. (2014), The Planetary and Lunar EphemeridesDE430 and DE431

J. G. Williams et al. (2013), DE430 Lunar Orbit, Physical Librations, and Surface Coordinates, DE430 Lunar Ephemeris and Orientation

FrameTransformations.Orient.DCM_MOON_PA440_TO_ME421 — Constant

DCM_MOON_PA440_TO_ME421

DCM for the rotation from the Moon Principal Axis 440 (PA440) to the Moon Mean Earth/Mean Rotation DE421 (ME421) axes.

References

Park, S. R. et al. (2021), The JPL Planetary and Lunar Ephemerides DE440 and DE441, DOI: 10.3847/1538-3881/abd414

FrameTransformations.Orient.IERS_EOP — Constant

IERS_EOP

Earth orientation parameters: x/y pole, UT1-UTC, LOD, dX, dY (smoothed values at 1-day intervals) with respect to IAU 2006/2000A precession-nutation model and consistent with ITRF2014.

EOP 14 C04 is updated two times per week.

Here the files are downloaded using the RemoteFile package with weekly updates.

See also: get_iers_eop

FrameTransformations.Orient.iau1980 — Constant

iau1980

The singleton instance of type IAU1980, representing the IAU 1980 family of models.

FrameTransformations.Orient.iau2000a — Constant

iau2000a

The singleton instance of type IAU2006a, representing the IAU 2000B family of models.

FrameTransformations.Orient.iau2000b — Constant

iau2000b

The singleton instance of type IAU2006a, representing the IAU 2000B family of models.

FrameTransformations.Orient.iau2006a — Constant

iau2006a

The singleton instance of type IAU2006a, representing the IAU 2006A family of models.

FrameTransformations.Orient.iau2006b — Constant

iau2006b

The singleton instance of type IAU2006B, representing the IAU 2006B family of models.

Note

This is not an official IERS model.