Utils

API

Types

FrameTransformations.Utils.TPC — Type

TPC <: JSMDInterfaces.FilesIO.AbstractFile

A type representing TPC files.

Functions

FrameTransformations.Utils._3angles_to_δdcm — Method

_3angles_to_δdcm(θ, rot_seq::Symbol)

Compute the time derivative of the DCM with all the angles stored in a single vector to optimise computations.

FrameTransformations.Utils._3angles_to_δ²dcm — Method

_3angles_to_δ²dcm(θ, rot_seq::Symbol)

Compute the 2nd time derivative of the DCM with all the angles stored in a single vector to optimise computations.

FrameTransformations.Utils.angle_to_δdcm — Function

angle_to_δdcm(θ[, ϕ[, γ]], rot_seq::Symbol = :ZYX)

Compute the derivative of the direction cosine matrix that perform a set of rotations (θ, ϕ, γ) about the coordinate axes in rot_seq. Each rotation input must be an indexable objected which includes the angle and its first time derivative.

The rotation sequence is defined by a Symbol specifing the rotation axes. The possible values depend on the number of rotations as follows:

1 rotation (θ₁): :X, :Y, or :Z.

2 rotations (θ₁, θ₂): :XY, :XZ, :YX, :YZ, :ZX, or :ZY.

3 rotations (θ₁, θ₂, θ₃): :XYX, XYZ, :XZX, :XZY, :YXY, :YXZ, :YZX, :YZY, :ZXY, :ZXZ, :ZYX, or :ZYZ

Note

This function assigns dcm = A3 * A2 * A1 in which Ai is the DCM related with the i-th rotation, i Є [1,2,3]. If the i-th rotation is not specified, then Ai = I.

See also

See also angle_to_δ²dcm

FrameTransformations.Utils.angle_to_δ²dcm — Function

angle_to_δ²dcm(θ[, ϕ[, γ]], rot_seq::Symbol = :ZYX)

Compute the second order time derivative of the direction cosine matrix that perform a set of rotations (θ, ϕ, γ) about the coordinate axes in rot_seq. Each rotation input must be an indexable objected which includes the angle and its first and second order time derivatives.

The rotation sequence is defined by a Symbol specifing the rotation axes. The possible values depend on the number of rotations as follows:

1 rotation (θ₁): :X, :Y, or :Z.

2 rotations (θ₁, θ₂): :XY, :XZ, :YX, :YZ, :ZX, or :ZY.

3 rotations (θ₁, θ₂, θ₃): :XYX, XYZ, :XZX, :XZY, :YXY, :YXZ, :YZX, :YZY, :ZXY, :ZXZ, :ZYX, or :ZYZ

Note

This function assigns dcm = A3 * A2 * A1 in which Ai is the DCM related with the i-th rotation, i Є [1,2,3]. If the i-th rotation is not specified, then Ai = I.

See also

See also angle_to_δdcm

FrameTransformations.Utils.angle_to_δ³dcm — Function

angle_to_δ³dcm(θ::Number[, ϕ::Number[, γ::Number]], rot_seq::Symbol = :Z)

Compute the second order time derivative of the direction cosine matrix that perform a set of rotations (θ, ϕ, γ) about the coordinate axes in rot_seq. Each rotation input must be an indexable objected which includes the angle and its time derivatives up to order 3 (jerk).

The rotation sequence is defined by a Symbol specifing the rotation axes. The possible values depend on the number of rotations as follows:

1 rotation (θ₁): :X, :Y, or :Z.

2 rotations (θ₁, θ₂): :XY, :XZ, :YX, :YZ, :ZX, or :ZY.

3 rotations (θ₁, θ₂, θ₃): :XYX, XYZ, :XZX, :XZY, :YXY, :YXZ, :YZX, :YZY, :ZXY, :ZXZ, :ZYX, or :ZYZ

Note

This function assigns dcm = A3 * A2 * A1 in which Ai is the DCM related with the i-th rotation, i Є [1,2,3]. If the i-th rotation is not specified, then Ai = I.

See also

See also angle_to_δdcm

FrameTransformations.Utils.cross12 — Method

cross12(x::AbstractVector, y::AbstractVector)

Compute the cross product between x and y and its 1st, 2nd and 3rd order time derivatives.

Notes

x and y must be 12-elements state vectors (position, velocity and acceleration)

FrameTransformations.Utils.cross3 — Method

cross3(x::AbstractVector, y::AbstractVector)

Compute the cross product between x and y, considering only their first 3 elements.

FrameTransformations.Utils.cross6 — Method

cross6(x::AbstractVector, y::AbstractVector)

Compute the cross product between x and y and its time derivative.

Notes

x and y must be 6-elements state vectors, with the last elements of each vector representing the time derivatives of the first three.

FrameTransformations.Utils.cross9 — Method

cross9(x::AbstractVector, y::AbstractVector)

Compute the cross product between x and y and its 1st and 2nd-order time derivatives.

Notes

x and y must be 9-elements state vectors (position, velocity and acceleration)

FrameTransformations.Utils.geoc2pos — Method

geoc2pos(r::Number, λ::Number, ϕ::Number)
geoc2pos(geoc::AbstractArray)

Transform geocentric coordinates in a cartesian position vector, given the longitude λ, the geocentric latitude ϕ and the radius r.

FrameTransformations.Utils.geoc2pv — Method

geoc2pv(geoc::AbstractVector)

Transform a spherical geocentric 6-elements state vector (radius, longitude, geocentric latitude and their derivatives) into a cartesian 6-elements vector (position and velocity).

FrameTransformations.Utils.geod2pos — Method

geod2pos(h::Number, λ::Number, ϕ::Number, R::Number, f::Number)

Transform longitude λ, geodetic latitude ϕ and altitude over the reference ellipsoid to a cartesian position vector, given the reference radius R and the flattening f.

References

Vallado, D. A (2013). Fundamentals of Astrodynamics and Applications, Microcosm Press, Hawthorn, CA, USA.

FrameTransformations.Utils.normalize — Method

normalize(v::AbstractVector)

Normalise the vector v.

FrameTransformations.Utils.pos2geoc — Method

pos2geoc(pos::AbstractVector)

Transform a cartesian 3-elements position vector pos into radius, longitude and geocentric latitude, respectively.

FrameTransformations.Utils.pos2geod — Function

pos2geod(pos::AbstratVector, R::Number, f::Number, toll::Number=1e-12)

Transform a cartesian 3-elements position vector pos into longitude, geodetic latitude and altitude over the reference ellipsoid with radius R and flattening f.

References

Vallado, D. A (2013). Fundamentals of Astrodynamics and Applications, Microcosm Press, Hawthorn, CA, USA.

FrameTransformations.Utils.pv2geoc — Method

pv2geoc(pv::AbstractVector)

Transform a cartesian 6-elements state vector (position and velocity) into radius, longitude, geocentric latitude and their derivatives, respectively.

FrameTransformations.Utils.skew — Method

skew(a)

Create a skew matrix from the first three elements of a.

FrameTransformations.Utils.δnormalize — Method

δnormalize(v::AbstractVector)

Compute the time derivative of a unit vector v.

FrameTransformations.Utils.δ²normalize — Method

δ²normalize(v::AbstractVector)

Compute the 2nd-order time derivative of a unit vector v.

FrameTransformations.Utils.δ³normalize — Method

δ³normalize(v::AbstractVector)

Compute the 3rd-order time derivative of a unit vector v.

JSMDInterfaces.FilesIO.load — Method

load(file::TPC{1})

Open a JPL ASCII .tpc file and parse its data in a dictionary.

JSMDInterfaces.FilesIO.load — Method

load(files::TPC)

Open a group of JPL ASCII .tpc files and parse their data in a dictionary.

Macros

Constants

FrameTransformations.Utils.light_speed — Constant

light_speed

Official light speed constant value used in CSPICE