Utils's Low-level API

Geodesy

FrameTransformations.Utils.pv2geoc — Function

pv2geoc(pv::AbstractVector)

Transform a cartesian 6-elements state vector (position and velocity) into radius, longitude, geocentric latitude and their derivatives, respectively.

FrameTransformations.Utils.geoc2pv — Function

geoc2pv(geoc::AbstractVector)

Transform a spherical geocentric 6-elements state vector (radius, longitude, geocentric latitude and their derivatives) into a cartesian 6-elements vector (position and velocity).

Vectors

FrameTransformations.Utils.cross3 — Function

cross3(x::AbstractVector, y::AbstractVector)

Compute the cross product between x and y, considering only their first 3 elements.

FrameTransformations.Utils.cross6 — Function

cross6(x::AbstractVector, y::AbstractVector)

Compute the cross product between x and y and its time derivative.

Notes

x and y must be 6-elements state vectors, with the last elements of each vector representing the time derivatives of the first three.

FrameTransformations.Utils.cross9 — Function

cross9(x::AbstractVector, y::AbstractVector)

Compute the cross product between x and y and its 1st and 2nd-order time derivatives.

Notes

x and y must be 9-elements state vectors (position, velocity and acceleration)

FrameTransformations.Utils.cross12 — Function

cross12(x::AbstractVector, y::AbstractVector)

Compute the cross product between x and y and its 1st, 2nd and 3rd order time derivatives.

Notes

x and y must be 12-elements state vectors (position, velocity and acceleration)

FrameTransformations.Utils.normalize — Function

normalize(v::AbstractVector)

Normalise the vector v.

FrameTransformations.Utils.δnormalize — Function

δnormalize(v::AbstractVector)

Compute the time derivative of a unit vector v.

FrameTransformations.Utils.δ²normalize — Function

δ²normalize(v::AbstractVector)

Compute the 2nd-order time derivative of a unit vector v.

FrameTransformations.Utils.δ³normalize — Function

δ³normalize(v::AbstractVector)

Compute the 3rd-order time derivative of a unit vector v.

Rotations

FrameTransformations.Utils.angle_to_δdcm — Function

angle_to_δdcm(θ[, ϕ[, γ]], rot_seq::Symbol = :ZYX)

Compute the derivative of the direction cosine matrix that perform a set of rotations (θ, ϕ, γ) about the coordinate axes in rot_seq. Each rotation input must be an indexable objected which includes the angle and its first time derivative.

The rotation sequence is defined by a Symbol specifing the rotation axes. The possible values depend on the number of rotations as follows:

1 rotation (θ₁): :X, :Y, or :Z.

2 rotations (θ₁, θ₂): :XY, :XZ, :YX, :YZ, :ZX, or :ZY.

3 rotations (θ₁, θ₂, θ₃): :XYX, XYZ, :XZX, :XZY, :YXY, :YXZ, :YZX, :YZY, :ZXY, :ZXZ, :ZYX, or :ZYZ

Note

This function assigns dcm = A3 * A2 * A1 in which Ai is the DCM related with the i-th rotation, i Є [1,2,3]. If the i-th rotation is not specified, then Ai = I.

See also

See also angle_to_δ²dcm

FrameTransformations.Utils.angle_to_δ²dcm — Function

angle_to_δ²dcm(θ[, ϕ[, γ]], rot_seq::Symbol = :ZYX)

Compute the second order time derivative of the direction cosine matrix that perform a set of rotations (θ, ϕ, γ) about the coordinate axes in rot_seq. Each rotation input must be an indexable objected which includes the angle and its first and second order time derivatives.

The rotation sequence is defined by a Symbol specifing the rotation axes. The possible values depend on the number of rotations as follows:

1 rotation (θ₁): :X, :Y, or :Z.

2 rotations (θ₁, θ₂): :XY, :XZ, :YX, :YZ, :ZX, or :ZY.

3 rotations (θ₁, θ₂, θ₃): :XYX, XYZ, :XZX, :XZY, :YXY, :YXZ, :YZX, :YZY, :ZXY, :ZXZ, :ZYX, or :ZYZ

Note

This function assigns dcm = A3 * A2 * A1 in which Ai is the DCM related with the i-th rotation, i Є [1,2,3]. If the i-th rotation is not specified, then Ai = I.

See also

See also angle_to_δdcm

FrameTransformations.Utils.angle_to_δ³dcm — Function

angle_to_δ³dcm(θ::Number[, ϕ::Number[, γ::Number]], rot_seq::Symbol = :Z)

Compute the second order time derivative of the direction cosine matrix that perform a set of rotations (θ, ϕ, γ) about the coordinate axes in rot_seq. Each rotation input must be an indexable objected which includes the angle and its time derivatives up to order 3 (jerk).

The rotation sequence is defined by a Symbol specifing the rotation axes. The possible values depend on the number of rotations as follows:

1 rotation (θ₁): :X, :Y, or :Z.

2 rotations (θ₁, θ₂): :XY, :XZ, :YX, :YZ, :ZX, or :ZY.

3 rotations (θ₁, θ₂, θ₃): :XYX, XYZ, :XZX, :XZY, :YXY, :YXZ, :YZX, :YZY, :ZXY, :ZXZ, :ZYX, or :ZYZ

Note

This function assigns dcm = A3 * A2 * A1 in which Ai is the DCM related with the i-th rotation, i Є [1,2,3]. If the i-th rotation is not specified, then Ai = I.

See also

See also angle_to_δdcm

FrameTransformations.Utils._3angles_to_δdcm — Function

_3angles_to_δdcm(θ, rot_seq::Symbol)

Compute the time derivative of the DCM with all the angles stored in a single vector to optimise computations.

FrameTransformations.Utils._3angles_to_δ²dcm — Function

_3angles_to_δ²dcm(θ, rot_seq::Symbol)

Compute the 2nd time derivative of the DCM with all the angles stored in a single vector to optimise computations.

FrameTransformations.Utils._3angles_to_δ³dcm — Function

_3angles_to_δ³dcm(θ, rot_seq::Symbol)

Compute the 3rd order time derivative of the DCM with all the angles stored in a single vector to optimise computations.

Constants

FrameTransformations.Utils.light_speed — Constant

light_speed

Official light speed constant value used in CSPICE