Orient's Public Documentation

IAU Models

This is a list of the supported IAU models and their approximations that can be used to select the desired procedure for the computation of the ITRF to GCRF rotation.

FrameTransformations.Orient.iau2000a — Constant

iau2000a

The singleton instance of type IAU2006a, representing the IAU 2000B family of models.

FrameTransformations.Orient.iau2000b — Constant

iau2000b

The singleton instance of type IAU2006a, representing the IAU 2000B family of models.

FrameTransformations.Orient.iau2006a — Constant

iau2006a

The singleton instance of type IAU2006a, representing the IAU 2006A family of models.

FrameTransformations.Orient.iau2006b — Constant

iau2006b

The singleton instance of type IAU2006B, representing the IAU 2006B family of models.

Note

This is not an official IERS model.

FrameTransformations.Orient.CPNc — Constant

CPNc

The singleton instance of type CPNC, representing the concise CPNc from Capitaine & Wallace, Concise CIO based precession-nutation formulations, (2008). This model truncates the X, Y series to deliver an accuracy of few mas.

FrameTransformations.Orient.CPNd — Constant

CPNd

The singleton instance of type CPND, representing the concise CPNd from Capitaine & Wallace, Concise CIO based precession-nutation formulations, (2008). This model truncates the X, Y series to deliver an accuracy of few arcseconds.

EOP Data

FrameTransformations.Orient.init_eop — Function

init_eop(filename)

Initialize Earth Orientation Parameters (EOP) from file.

Warn

This function must be called to initialize the EOP data used by frames, in case Earth-associated frames are used.

Warn

This function accept only .eop.dat files. Please use prepare_eop to transform IERS EOP files in this format.

FrameTransformations.Orient.prepare_eop — Function

prepare_eop(iers_file::AbstractString, output_filename::AbstractString="iau2000a")

Prepare Earth Orientation Parameters (EOP) data from IERS EOP C04 files to JSMD's eop.dat convenience format. The output_filename should not include the file extension, which is automatically added by this function.

# Save a new file called: test.eop.dat
prepare_eop("input.csv", "test")

FrameTransformations.Orient.eop_data_filename — Function

eop_data_filename()

Get loaded Earth Orientation Parameters (EOP) filename.

FrameTransformations.Orient.IERS_EOP — Constant

IERS_EOP

Earth Orientation Parameters interpolators: x/y pole, UT1-UTC, LOD, dX, dY (smoothed values at 1-day intervals) with respect to IAU 2006/2000A precession-nutation model and consistent with ITRF2014.

See also: EOPInterpolator

IERS Transformations

FrameTransformations.Orient.orient_rot3_itrf_to_gcrf — Function

orient_rot3_itrf_to_gcrf(m::IAUModel, t::Number)

Compute the rotation matrix from ITRF to GCRF at time t expressed as TT seconds since J2000, according to the IAU Model m, as follows:

IAU2000A: the pole coordinates (xₚ, yₚ) and the free-core nutation and time corrections to the CIP coordinates (dX, dY) are interpolated from the latest released IERS EOP data. The precession-nutation matrix is computed using the full IAU 2000A model.

IAU2000B: only the pole coordinates (xₚ, yₚ) are interpolated from the latest EOP data. The Free Core Nutation (FCN) corrections dX, dY are neglected. The precession-nutation matrix is computed following the IAU 2000 model but with truncated expressions for the nutation corrections.

IAU2006A: the pole coordinates (xₚ, yₚ) and the free-core nutation and time corrections to the CIP coordinates (dX, dY) are interpolated from the latest released IERS EOP data. The precession-nutation matrix is computed using the full IAU 2006/2000A model.

IAU2006B: only the pole coordinates (xₚ, yₚ) are interpolated from the latest EOP data. The Free Core Nutation (FCN) corrections dX, dY are neglected. The precession-nutation matrix is computed following the IAU 2006A model but with truncated expressions for the nutation corrections.

CPNc: a concise model with a cut-off at 2.5 mas of the X and Y series, delivering a worst-case accuracy of about 15 mas between 1995-2050. It does not take into account the Free Core Nutation (~0.2 mas).

CPNd: an extremely concise formulation with an accuracy of about 1 arcsec between 1995 and 2050. It neglects polar-motion (~0.25 arcsec), the FCN corrections and the CIO locator.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

orient_rot3_itrf_to_gcrf(m::IAUModel, tt, ut1, xₚ, yₚ, dX=0.0, dY=0.0)

Compute the rotation matrix from ITRF to GCRF according to the IAU Model m, at time tt and ut1 expressed in TT seconds and UT1 days since J2000, respectively.

This function has been implemented for IAU2000, IAU2006 and CPN models.

Note

All the input quantities xₚ, yₚ, dX and dY must be expressed in radians

FrameTransformations.Orient.orient_rot6_itrf_to_gcrf — Function

orient_rot6_itrf_to_gcrf(m::IAUModel, t::Number)

Compute the rotation matrix from ITRF to GCRF and its derivative at time t expressed as TT seconds since J2000, according to the the IAU Model m.

FrameTransformations.Orient.orient_rot9_itrf_to_gcrf — Function

orient_rot9_itrf_to_gcrf(m::IAUModel, t::Number)

Compute the rotation matrix from ITRF to GCRF and its time derivatives up to order 2 at time t expressed as TT seconds since J2000, according to the IAU Model m.

FrameTransformations.Orient.orient_rot12_itrf_to_gcrf — Function

orient_rot12_itrf_to_gcrf(m::IAUModel, t::Number)

Compute the rotation matrix from ITRF to GCRF and its time derivatives up to order 3 at time t expressed as TT seconds since J2000, according to the IAU Model m.

FrameTransformations.Orient.orient_bias_precession — Function

orient_bias_precession(m::IAUModel, t::Number)

Form the precession-frame bias (PB) matrix that transforms vectors from the GCRS to the mean of date, following the IAU Model m at time t expressed as TT Julian centuries since J2000.

The function has been implemented for the IAU2000 and IAU2006 models.

References:

IAU: Trans. International Astronomical Union, Vol. XXIVB; Proc. 24th General Assembly, Manchester, UK. Resolutions B1.3, B1.6. (2000)

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA pmat06 function.

ERFA pmat00 function.

FrameTransformations.Orient.orient_bias_precession_nutation — Function

orient_bias_precession_nutation(m::IAUModel, t::Number)

Compute the equinox-based bias-precession-nutation matrix using the IAU Model m procedures at time t expressed in TT Julian centuries since J2000.

The function has been implemented for the IAU2000 and IAU2006 models.

Note

The computed matrix rotates a vector from the GCRS to the true equatorial triad of date.

References

ERFA pn06 function

ERFA pn00 function

FrameTransformations.Orient.orient_gast — Function

orient_gast(m::IAUModel, t::Number)

Compute the Greenwich Apparent Sidereal Time (GAST), in radians, given time t as TT Julian centuries since J2000 according to the IAU Model m.

Note

For the IAU2000B model, as an approximation ERFA uses UT1 instead of TDB (or TT) to compute the precession component of GMST and the equation of the equinoxes. This approximation is not performed in this framework.

Note

This function computes the Earth Rotation Angle (ERA) by automatically converting TT to UT1. Lower-level interfaces are also available to minimise the number of repeated computations.

See also

See also orient_gmst.

orient_gast(m::IAUModel, t::Number, θ::Number)

Compute the Greenwich Apparent Sidereal Time (GAST), in radians, given time t as TT Julian centuries since J2000 and the Earth Rotation Angle (ERA) θ, in radians, according to the IAU Model m.

The function has been implemented for the IAU2000 and IAU2006 models.

References

ERFA gst00a routine.

ERFA gst06a routine.

FrameTransformations.Orient.orient_gmst — Function

orient_gmst(::IAUModel, t::Number)

Compute the Greenwich Mean Sidereal Time (GMST), in radians, according to the IAU Model m, given time tt expressed in TT Julian centuries since J2000.

Note

This function is valid for the IAU2000 and IAU2006 models only.

Note

This function computes the Earth Rotation Angle (ERA) by automatically converting TT to UT1. Lower-level interfaces are also available to minimise the number of repeated computations.

See also

See also orient_gast.

orient_gmst(m::IAU1980Model, ut1::Number)

Compute the Greenwich Mean Sidereal Time (GMST), in radians, according to the IAU 1980 models, given time ut1 expressed in UT1 Julian centuries since J2000.

References

ERFA gmst82 routine.

orient_gmst(m::IAUModel, tt::Number, θ::Number)

Compute the Greenwich Mean Sidereal Time (GMST), in radians, according to the IAU Model m, given the Earth Rotation Angle (ERA) θ, in radians and the time tt expressed in TT Julian centuries since J2000.

The function has been implemented for the IAU2000 and IAU2006 models.

References

ERFA gmst00 routine.

ERFA gmst06 routine.

FrameTransformations.Orient.orient_nutation — Function

orient_nutation(m::IAUModel, t::Number)

Compute the nutation components in longitude and obliquity for the IAU Model m, in radians, at time t expressed in TT Julian Centuries since J2000.

Notes

Due to their theoretical basis, the original developments required t expressed as TDB. However, in practice, it is usually more convenient to use Terrestrial Time (TT) as it makes no significant differences (< 0.01 μas) in the final result.

For the IAU 2006A model, the function strictly follows the SOFA implementation. It first computes the IAU 2000A nutation, then applies adjustments for the consequences of the change in obliquity from the IAU 1980 ecliptic to the IAU 2006 ecliptic and (ii) for the secular variation in the Earth's dynamical form factor J2. These corrections ensure that the IAU 2000A nutation is consistent with the IAU 2006 precession model. Please note that the coefficients available on the IERS tables already include those corrections, and are retrieved by multiplying the amplitudes of the SOFA nutation in longitude coefficients by 1.00000047.

The computation of the free-core nutation and time dependent effects are excluded from this model. To achieve the < 1μas accuracy with the IAU 2006/2000 A precession-nutation models, such effects must be included a-posteriori (through dX and dY) using the IERS EOP data.

For the IAU 2000B model, the nutation series is truncated from nearly 1400 terms to only 77, yet it still delivers results of 1 mas accuracy at present epochs. In particular, it delivers a pole accurate to 1 mas from 1900 to 2100 (only very occasionally just outside 1 mas). The coefficients are taken from SOFA's implementation, which slighlty differ from those reported in McCarthy and Luzum (2003). Comparisons with IAU 2006A show that the SOFA version between 1995 and 2050 delivers 0.283 mas RMSE (0.994 mas in the worst case), whereas the IERS Conventions website version delivers 0.312 RMSE (1.125 mas in the worst case).

The IAU 2000B model includes constant planetary bias terms that compensate for long-period nutations. These amplitudes used in this implementation are optimised for a rigorous method, where frame bias, precession and nutation are applied separately and in that order (see SOFA's documentation for further insights).

A simplified version of the Fundamental Arguments, taken from Simon et al (1994) is exploited for IAU2000B as the error introduced is below the model accuracy ( ~0.1 mas).

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Simon, J. et al., (1994), Numerical expressions for precession formulae and mean elements for the Moon and the planets.

ERFA nut06a and nut00b functions

FrameTransformations.Orient.orient_obliquity — Function

orient_obliquity(m::IAUModel, t::Number)

Compute the mean obliquity of the ecliptic at epoch, in radians, at time t expressed in TT Julian centuries since J2000.

Note

This function is implemented only for IAU1980 and IAU2006 models. IAU 2000 Models implement proper precession-rate corrections to the IAU1980 mean obliquity.

References

ERFA obl80 and obl06 functions.

Moon Transformations

FrameTransformations.Orient.orient_rot3_icrf_to_pa421 — Function

orient_rot3_icrf_to_pa421(eph::AbstractEphemerisProvider, t::Number)

Compute the rotation matrix from the ICRF to the DE421 Moon's Principal Axes at the given input time t, expressed in seconds since J2000.

Warning

This function is not optimised for performance (it allocates!). The user is suggested to retrieve the Principal axes orientation using the dedicated Frame System functions.

orient_rot3_icrf_to_pa421(eph::AbstractEphemerisProvider, ep::Epoch)

Compute the rotation matrix from the ICRF to the DE421 Moon's Principal Axes at the input epoch ep.

FrameTransformations.Orient.orient_rot3_icrf_to_pa440 — Function

orient_rot3_icrf_to_pa440(eph::AbstractEphemerisProvider, t::Number)

Compute the rotation matrix from the ICRF to the DE440 Moon's Principal Axes at the given input time t, expressed in seconds since J2000.

Warning

This function is not optimised for performance (it allocates!). The user is suggested to retrieve the Principal axes orientation using the dedicated Frame System functions.

orient_rot3_icrf_to_pa440(eph::AbstractEphemerisProvider, ep::Epoch)

Compute the rotation matrix from the ICRF to the DE440 Moon's Principal Axes at the input epoch ep.

Geodesy

FrameTransformations.Orient.geoc2pos — Function

geoc2pos(r::Number, λ::Number, ϕ::Number)
geoc2pos(geoc::AbstractArray)

Transform geocentric coordinates in a cartesian position vector, given the longitude λ, the geocentric latitude ϕ and the radius r.

FrameTransformations.Orient.pos2geoc — Function

pos2geoc(pos::AbstractVector)

Transform a cartesian 3-elements position vector pos into radius, longitude and geocentric latitude, respectively.

FrameTransformations.Orient.geod2pos — Function

geod2pos(h::Number, λ::Number, ϕ::Number, R::Number, f::Number)

Transform longitude λ, geodetic latitude ϕ and altitude over the reference ellipsoid to a cartesian position vector, given the reference radius R and the flattening f.

References

Vallado, D. A (2013). Fundamentals of Astrodynamics and Applications, Microcosm Press, Hawthorn, CA, USA.

FrameTransformations.Orient.pos2geod — Function

pos2geod(pos::AbstratVector, R::Number, f::Number, toll::Number=1e-12)

Transform a cartesian 3-elements position vector pos into longitude, geodetic latitude and altitude over the reference ellipsoid with radius R and flattening f.

References

Vallado, D. A (2013). Fundamentals of Astrodynamics and Applications, Microcosm Press, Hawthorn, CA, USA.

Axes ID

This is a list of NAIF IDs for standard axes that are used in astrodynamic applications.

Note

Although they are listed in the public documentation section, these IDs are not directly exported by the package.

FrameTransformations.Orient.AXESID_B1950 — Constant

AXESID_B1950

NAIF Axes ID for the Mean Equator and Dynamical Equinox of the Besselian year 1950.

FrameTransformations.Orient.AXESID_ECLIPB1950 — Constant

AXESID_ECLIPB1950

NAIF Axes ID for the Mean Ecliptic Equinox of B1950.

FrameTransformations.Orient.AXESID_ECLIPJ2000 — Constant

AXESID_ECLIPJ2000

NAIF Axes ID for the Mean Ecliptic Equinox of J2000 (ECLIPJ2000)

FrameTransformations.Orient.AXESID_FK4 — Constant

AXESID_FK4

NAIF Axes ID for the FK4 reference frame.

FrameTransformations.Orient.AXESID_GALACTIC — Constant

AXESID_GALACTIC

NAIF Axes ID for the Galactic System II reference frame.

FrameTransformations.Orient.AXESID_GCRF — Constant

AXESID_GCRF

Axes ID for the Geocentric Celestial Reference Frame (GCRFF)

Note

Although the ICRF and GCRF axes are identical, they are based upon a different timescale. A different ID is here assigned to provide a robust way of distinguishing between the two. 23 has been chosen because it is one the unassigned axes ID among the built-in SPICE frames.

FrameTransformations.Orient.AXESID_ICRF — Constant

AXESID_ICRF

NAIF Axes ID for the International Celestial Reference Frame (ICRF)

FrameTransformations.Orient.AXESID_ITRF — Constant

AXESID_ITRF

NAIF Axes ID for the International Terrestrial Reference Frame (ITRF)

Note

This ID is based upon the ID used to refer to the ITRF93 in NAIF's high-accuracy Earth rotation model PCK kernels.

FrameTransformations.Orient.AXESID_MEME2000 — Constant

AXESID_MEME2000

Axes ID for the Mean Dynamical Equator and Equinox of J2000.0.

Note

In SPICE the J2000 and ICRF axes are considered equal, thus there exist no specific NAIF ID for the MEME2000 axes. 22 has been chosen because it is the first unassigned axes ID among the built-in SPICE frames.

FrameTransformations.Orient.AXESID_MOONME_DE421 — Constant

AXESID_MOONME_DE421

NAIF axes id for the DE421 Moon Mean Earth/Mean Rotation axes (ME421).

FrameTransformations.Orient.AXESID_MOONPA_DE421 — Constant

AXESID_MOONPA_DE421

NAIF axes id for the DE421 Moon Principal Axes (PA421).

FrameTransformations.Orient.AXESID_MOONPA_DE440 — Constant

AXESID_MOONPA_DE440

NAIF Axes id for the DE440 Moon Principal Axes (PA440).

Default Rotation Matrices

Note

Although they are listed in the public documentation section, the rotation matrices of some older frames (e.g., B1950, FK4 and GALACTIC) are not exported by the package.

FrameTransformations.Orient.DCM_B1950_TO_ECLIPB1950 — Constant

DCM_B1950_TO_ECLIPB1950

DCM for the rotation from the Mean Equator and Dynamical Equinox of B1950 (MEMEB1950) to the Mean Ecliptic Equinox of B1950. This corresponds to the transformation B1950 -> ECLIPB1950 in the SPICE toolkit, and uses the mean obliquity of the ecliptic from the IAU 1976 theory.

References

SPICE Library

FrameTransformations.Orient.DCM_B1950_TO_FK4 — Constant

DCM_B1950_TO_FK4

DCM for the rotation from the Mean Equator and Dynamical Equinox of B1950 (MEMEB1950) ot the FK4 reference frame.

Note

The FK4 reference frame is obtained from the B1950 frame by applying the equinox offset determined by Fricke.

References

SPICE Library

FrameTransformations.Orient.DCM_ICRF_TO_ECLIPJ2000 — Constant

DCM_ICRF_TO_ECLIPJ2000

DCM for the rotation from the International Celestial Reference Frame (ICRF) to the Mean Ecliptic Equinox of J2000 (ECLIPJ2000).

FrameTransformations.Orient.DCM_ICRF_TO_MEME2000 — Constant

DCM_ICRF_TO_MEME2000

DCM for the rotation from the International Celestial Reference Frame (ICRF) and the Mean Equator and Equinox of J2000.0 (MEME2000). This corresponds to the J2000 frame in the SPICE toolkit.

Note

The frame bias is here computed using the IAU 2006 Precession model, similarly to ESA's GODOT. Some other software libraries, such as Orekit, use the frame bias of the IAU 2000 precession model. The two definitions differ of about 1 arcsecond.

Moreover, according to Hilton there are multiple possibilities to define the proper rotation between the ICRS and the MEME2000. The transformation implemented here correspond to Eq. 6 using the parameters in Table 3, line 1 (RIERS).

References

Hilton, James L., and Catherine Y. Hohenkerk. – Rotation matrix from the mean dynamical equator and equinox at J2000. 0 to the ICRS. – Astronomy & Astrophysics 513.2 (2004): 765-770. DOI: 10.1051/0004-6361:20031552

SOFA docs

FrameTransformations.Orient.DCM_FK4_TO_GALACTIC — Constant

DCM_FK4_TO_GALACTIC

DCM for the rotation from the FK4 frame to the Galactic System II reference frame.

Note

As the SPICE toolkit, we assume that this rotation is derived from the FK4 frame.

References

SPICE Library

FrameTransformations.Orient.DCM_MEME2000_TO_ECLIPJ2000 — Constant

DCM_MEME2000_TO_ECLIPJ2000

DCM for the rotation from the Mean Equator and Equinox of J2000 (MEME2000) to the Mean Ecliptic Equinox. This corresponds to the transformation J2000 -> ECLIPJ2000 in the SPICE toolkit, and uses the mean obliquity of the ecliptic from the IAU 1976 theory.

FrameTransformations.Orient.DCM_MEME2000_TO_B1950 — Constant

DCM_MEME2000_TO_B1950

DCM for the rotation from the Mean Equator and Equinox of J2000.0 (MEME2000) to the Mean Equator and Dynamical Equinox of B1950 (B1950).

Note

This rotation is obtained by precessing the J2000 frame backwards from Julian year 2000 to Besselian year 1950, using the 1976 IAU precession model. The rotation values are taken from the SPICE toolkit.

References

SPICE Library

FrameTransformations.Orient.DCM_MOON_PA421_TO_ME421 — Constant

DCM_MOON_PA421_TO_ME421

DCM for the rotation from the Moon Principal Axis 421 (PA421) to the Moon Mean Earth/Mean Rotation DE421 (ME421) axes.

References

J. G. Williams et al. (2008), DE421 Lunar Orbit, Physical Librations, and Surface Coordinates, DE421 Lunar Ephemeris and Orientation

FrameTransformations.Orient.DCM_MOON_PA430_TO_ME421 — Constant

DCM_MOON_PA430_TO_ME421

DCM for the rotation from the Moon Principal Axis 430 (PA430) to the Moon Mean Earth/Mean Rotation DE421 (ME421) axes.

References

Folkner M. William et al. (2014), The Planetary and Lunar EphemeridesDE430 and DE431

J. G. Williams et al. (2013), DE430 Lunar Orbit, Physical Librations, and Surface Coordinates, DE430 Lunar Ephemeris and Orientation

FrameTransformations.Orient.DCM_MOON_PA430_TO_ME430 — Constant

DCM_MOON_PA430_TO_ME430

DCM for the rotation from the Moon Principal Axis 430 (PA430) to the Moon Mean Earth/Mean Rotation DE430 (ME430) axes.

References

Folkner M. William et al. (2014), The Planetary and Lunar EphemeridesDE430 and DE431

J. G. Williams et al. (2013), DE430 Lunar Orbit, Physical Librations, and Surface Coordinates, DE430 Lunar Ephemeris and Orientation

FrameTransformations.Orient.DCM_MOON_PA440_TO_ME421 — Constant

DCM_MOON_PA440_TO_ME421

DCM for the rotation from the Moon Principal Axis 440 (PA440) to the Moon Mean Earth/Mean Rotation DE421 (ME421) axes.

References

Park, S. R. et al. (2021), The JPL Planetary and Lunar Ephemerides DE440 and DE441, DOI: 10.3847/1538-3881/abd414