Orient's Low-level API

Although this routines are not meant to be used outside of the package, they are here documented to aid future developments or to help users that require specific computations.

Outdated IAU Models

FrameTransformations.Orient.iau1980 — Constant

iau1980

The singleton instance of type IAU1980, representing the IAU 1980 family of models.

Fundamental Arguments

FrameTransformations.Orient.FundamentalArguments — Type

FundamentalArguments{N <: Number}

Type storing the fundamental luni-solar and planetary arguments.

Fields

Mₐ – Mean anomaly of the Moon

Sₐ – Mean anomaly of the Sun

uₘ – Mean longitude of the Moon minus mean longitude of the ascending node F

Dₛ – Mean elongation of the Moon from the Sun

Ωₘ – Mean longitude of the Moon's ascending node

λ_Me – Mercury's mean heliocentric longitude

λ_Ve – Venus's mean heliocentric longitude

λ_Ea – Earth's mean heliocentric longitude

λ_Ma – Mars's mean heliocentric longitude

λ_Ju – Jupiter's mean heliocentric longitude

λ_Sa – Saturn's mean heliocentric longitude

λ_Ur – Uranus's mean heliocentric longitude

λ_Ne – Neptune's mean heliocentric longitude

pₐ – General precession in longitude

FrameTransformations.Orient.LuniSolarArguments — Function

LuniSolarArguments(t::Number, m::IAUModel)

Compute the fundamental (Delaunay) Luni-Solar arguments, in radians, associated to the desired IAU model m, at time t expressed in TDB Julian centuries since J2000.

The returned values depend on the input model as follows:

IAU2006A: the Delaunay expressions are taken from the IERS 2010 Conventions.

IAU2000B: the expressions are taken from Simon et al. (1994), following ERFA's implementation of nut00b.c

Outputs

Mₐ – Mean anomaly of the Moon

Sₐ – Mean anomaly of the Sun

uₘ – Mean longitude of the Moon minus mean longitude of the ascending node F

Dₛ – Mean elongation of the Moon from the Sun

Ωₘ – Mean longitude of the Moon's ascending node

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

Simon, J. et al., (1994), Numerical expressions for precession formulae and mean elements for the Moon and the planets.

ERFA software library

FrameTransformations.Orient.PlanetaryArguments — Function

PlanetaryArguments(t::Number)

Compute the fundamental planetary arguments and the general precession, in radians, at time t expressed in TDB Julian centuries since J2000.

Outputs

λ☿ – Mercury's mean heliocentric longitude.

λ♀ – Venus's mean heliocentric longitude.

λe – Earth's mean heliocentric longitude.

λ♂ – Mars's mean heliocentric longitude.

λ♃ – Jupiter's mean heliocentric longitude.

λ♄ – Saturn's mean heliocentric longitude.

λ⛢ – Uranus's mean heliocentric longitude.

λ♆ – Neptune's mean heliocentric longitude.

pλ – General precession in longitude.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

FrameTransformations.Orient.fa_mano_moon — Function

fa_mano_moon(t::Number)

Return the mean anomaly of the Moon Mₐ, in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mano_sun — Function

fa_mano_sun(t::Number)

Return the mean anomaly of the Sun Sₐ in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_moon — Function

fa_mlon_moon(t::Number)

Return the mean longitude of the Moon's ascending node Ω in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlat_moon — Function

fa_mlat_moon(t::Number)

Return the mean longitude of the Moon minus the mean longitude of the ascending node F in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_melo_moon — Function

fa_melo_moon(t::Number)

Return the mean elongation of the Moon from the Sun D in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_precession — Function

fa_precession(t::Number)

Return the general accumulated precession in longitude pₐ in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_mercury — Function

fa_mlon_mercury(t::Number)

Return the mean heliocentric longitude of Mercury in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_venus — Function

fa_mlon_venus(t::Number)

Return the mean heliocentric longitude of Venus in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_earth — Function

fa_mlon_earth(t::Number)

Return the mean heliocentric longitude of Earth in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_mars — Function

fa_mlon_mars(t::Number)

Return the mean heliocentric longitude of Mars in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_jupiter — Function

fa_mlon_jupiter(t::Number)

Return the mean heliocentric longitude of Jupiter in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_saturn — Function

fa_mlon_saturn(t::Number)

Return the mean heliocentric longitude of Saturn in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_uranus — Function

fa_mlon_uranus(t::Number)

Return the mean heliocentric longitude of Uranus in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

FrameTransformations.Orient.fa_mlon_neptune — Function

fa_mlon_neptune(t::Number)

Return the mean heliocentric longitude of Neptune in radians, at time t expressed in TDB Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012). The IERS Conventions (2010), IERS Technical Note No. 36

ERFA software library

Precession

FrameTransformations.Orient.frame_bias — Function

frame_bias(::IAU2000Model)

Compute the frame bias components of the IAU 2000 precession-nutation models, in radians.

Notes

The frame bias corrections in longitude and obliquity are required to correct for the offset between the GCRS pole and the mean J2000 pole. They define, with respect to the GCRS axes, a J2000 mean pole that is consistent with teh IAU 2000A precession-nutation model.

The function also returns an offset in right ascension taken from Chapront et al. (2002), necessary to completely describe the frame bias, but that is not part of the original IAU 2000A model.

References

ERFA software library

FrameTransformations.Orient.fw_angles — Function

fw_angles(m::IAU2006Model, t::Number)

Compute the precession angles in radians, following the IAU 2006 Fukushima-Williams 4-angle formulation at time t expressed in TT Julian centuries since J2000.

Outputs

γ – F-W 1st angle

ϕ – F-W 2nd angle

ψ – F-W 3rd angle

ε – F-W 4th angle

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library

FrameTransformations.Orient.fw_matrix — Function

fw_matrix(γ, ϕ, ψ, ε)

Form the Nutation-Precession-Bias (NPB) rotation matrix given the Fukushima-Williams angles, expressed in radians.

The present function can construct three different matrices depending on which angles are supplied as arguments:

NPB: To obtain the Nutation-Precession-Bias matrix, generate the four standard FW precession angles (̄γ, ̄ϕ, ̄ψ, ϵₐ) then generate the nutation components Δψ and Δϵ and add them to ̄ψ, ϵₐ. Finally, call the present functions using those four angles as arguments.

PB: To obtain the precession-frame bias matrix, generate the four standard FW precession angles and call the present function.

B: To obtain the frame bias matrix, generate the four standard FW precession angles at date J2000.0 and call this function.

The remaining nutation-only and precession only matrices can be obtained by combining these three appropriately.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library

FrameTransformations.Orient.precession_angles — Function

precession_angles(m::IAU1980Model, t::Number)

Compute the precession angles from Lieske et al. 1977 model, in radians, at time t expressed in TT Julian centuries since J2000.

References

ERFA software library

FrameTransformations.Orient.precession_rate — Function

precession_rate(m::IAU2000Model, t::Number)

Compute the precession-rate part of the IAU 2000 precession-nutation models, in radians, at time t expressed as TT Julian centuries since J2000.

References

ERFA software library

IERS Routines

FrameTransformations.Orient.bpn2xy — Function

bpn2xy(A::AbstractMatrix)

Compute the CIP X and Y coordinates from the bias-precession-nutation matrix, in radians.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

FrameTransformations.Orient.cip_coords — Function

cip_coords(m::IAUModel, t::Number)

Computes the CIP X, Y coordinates, in radians, according to the IAU model m at time t expressed in TT Julian Centuries since J2000.

This function has been implemented for the IAU2000, IAU2006 and the CPN models.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

FrameTransformations.Orient.cio_locator — Function

cio_locator(m::IAUModel, t::Number, x::Number, y::Number)

Compute the CIO Locator s in radians, according to the IAU Model m, given the CIP coordinates X and Y at time t expressed in TT Julian centuries since J2000

The function has been implemented for the IAU2000, IAU2006 and the CPN models.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

ERFA library

ERFA library

FrameTransformations.Orient.cip_motion — Function

cip_motion(m::IAUModel, t::Number, dx::Number=0.0, dy::Number=0.0)

Compute the CIRS to GCRS rotation matrix, according to the IAU Model m, at time t expressed in TT Julian centuries since J2000. Optional IERS corrections for free-core nutation and time depedented effects can be provided through dx and dy.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

FrameTransformations.Orient.earth_rotation_angle — Function

earth_rotation_angle(t::Number)

Compute the Earth Rotation Angle (ERA) in radians, i.e., the angle between the Celestial Intermediate Origin (CIO) and the Terrestrial Intermediate Origin (TIO) at time t expressed as UT1 days since J2000.

Note

The function uses the fractional UT1 date to gain additional precision in the computations (0.002737.. instead of 1.002737..)

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

ERFA library

FrameTransformations.Orient.earth_rotation_rate — Function

earth_rotation_rate()

Compute the nominal Earth angular velocity.

earth_rotation_rate(LOD::Number)

Compute the true angular velocity of the Earth accounting for the Length of the Day, i.e., the instantaneous rate of change of UT1 with respect to a uniform time scale.

FrameTransformations.Orient.era_rotm — Function

era_rotm(t::Number)

Compute the TIRS to CIRS Earth Rotation matrix, according to the IERS 2010 conventions at time t expressed as UT1 days since J2000.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

FrameTransformations.Orient.fw2xy — Function

fw2xy(ϵ::Number, ψ::Number, γ::Number, φ::Number)

Compute the CIP X and Y coordinates from Fukushima-Williams bias-precession-nutation angles, in radians.

Inputs

ϵ – F-W angle with IAU 2006A/B nutation corrections.

ψ – F-W angle with IAU 2006A/B nutation corrections.

γ – F-W angle

ϕ – F-W angle

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

FrameTransformations.Orient.tio_locator — Function

tio_locator(t::Number)

Compute the TIO locator s' at date, positioning the Terrestrial Intermediate Origin on the equator of the Celestial Intermediate Pole (CIP) at time t expressed as TT Julian centuries since J2000.

This function approximates the unpredictable motion of the TIO locator s' with its secular drift of ~0.47 μas/century.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Lambert, S. and Bizouard C. (2002), Positioning the Terrestrial Ephemeris Origin in the Terrestrial Reference Frame, DOI: 10.1051/0004-6361:20021139

FrameTransformations.Orient.polar_motion — Function

polar_motion(t::Number, xₚ::Number, yₚ::Number)

Compute the Polar Motion rotation matrix from ITRF to TIRS, according to the IERS 2010 Conventions, at time t expressed in TT Julian centuries since J2000. The function requires xp and yp, the Celestial Intermediate Pole (CIP) coordinates with respect to the International Celestial Reference Frame (ITFR).

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

FrameTransformations.Orient.xys2m — Function

xys2m(x::Number, y::Number, s::Number)

Compute the Intermediate-to-Celestial matrix given the CIP x, y' coordinates and the CIO locators`, all in radians.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library

EOP Data

FrameTransformations.Orient.EOPData — Type

EOPData{T}

EOP Data for IAU 2000A.

Note

Each field will be an AbstractInterpolation indexed by the Julian Day.

Fields

x, y: Polar motion with respect to the crust [arcsec].

UT1_UTC: Irregularities of the rotation angle [s].

LOD: Length of day offset [ms].

dX, dY: Celestial pole offsets referred to the model IAU2000A [milliarcsec].

FrameTransformations.Orient.get_iers_eop — Function

get_iers_eop(; force_download = false)

Download and parse the IERS EOP C04 data.

The files are downloaded using the RemoteFile package with weekly updates. Hence, if one desires to force a download before the scheduled time, then set the keyword force_download to true.

Note

The files will be downloaded from the default URL. If the user want to use another one, then use the specialized function get_iers_eop_IAU2000A

See also: get_iers_eop_IAU2000A

Returns

A structure EOPData with the interpolations of the EOP parameters. Notice that the interpolation indexing is set to Julian days since J2000.

FrameTransformations.Orient.get_iers_eop_IAU2000A — Function

get_iers_eop_iau_2000A(url::String="https://datacenter.iers.org/data/csv/finals2000A.all.csv"; 
    force_download = false)

Get the IERS EOP C04 IAU2000A data from the URL url.

If url is omitted, then it defaults to https://datacenter.iers.org/data/csv/finals2000A.all.csv. The file is downloaded using the RemoteFile package with weekly updates. Hence, if one desires to force a download before the scheduled time, then set the keyword force_download to true.

Note

The interpolation of every field in EOPData between two points in the grid is linear. If extrapolation is needed, then if will use the nearest value (flat extrapolation).

See also: get_iers_eop

Returns

The structure EOPData with the interpolations of the EOP parameters. Notice that the interpolation indexing is set to Julian days since J2000.

FrameTransformations.Orient.offset_utc2ut1 — Function

offset_utc2ut1(seconds)

Return the offset between UTC and UT1 in seconds.

Transformations

FrameTransformations.Orient.orient_rot3_icrf_to_mememod — Function

orient_rot3_icrf_to_mememod(t::Number)

Compute the rotation matrix from the International Celestial Reference Frame (ICRF) to the Mean Equinox Mean Equator of Date at time t, expressed in TT seconds since J2000.