Utils's Public Documentation

TPC

FrameTransformations.Utils.TPC — Type

TPC <: JSMDInterfaces.FilesIO.AbstractFile

A type representing TPC files.

JSMDInterfaces.FilesIO.load — Function

load(file::TPC{1})

Open a JPL ASCII .tpc file and parse its data in a dictionary.

load(files::TPC)

Open a group of JPL ASCII .tpc files and parse their data in a dictionary.

Math

FrameTransformations.Utils.skew — Function

skew(a)

Create a skew matrix from the first three elements of a.

Geodesy

FrameTransformations.Utils.geoc2pos — Function

geoc2pos(r::Number, λ::Number, ϕ::Number)
geoc2pos(geoc::AbstractArray)

Transform geocentric coordinates in a cartesian position vector, given the longitude λ, the geocentric latitude ϕ and the radius r.

FrameTransformations.Utils.pos2geoc — Function

pos2geoc(pos::AbstractVector)

Transform a cartesian 3-elements position vector pos into radius, longitude and geocentric latitude, respectively.

FrameTransformations.Utils.geod2pos — Function

geod2pos(h::Number, λ::Number, ϕ::Number, R::Number, f::Number)

Transform longitude λ, geodetic latitude ϕ and altitude over the reference ellipsoid to a cartesian position vector, given the reference radius R and the flattening f.

References

Vallado, D. A (2013). Fundamentals of Astrodynamics and Applications, Microcosm Press, Hawthorn, CA, USA.

FrameTransformations.Utils.pos2geod — Function

pos2geod(pos::AbstratVector, R::Number, f::Number, toll::Number=1e-12)

Transform a cartesian 3-elements position vector pos into longitude, geodetic latitude and altitude over the reference ellipsoid with radius R and flattening f.

References

Vallado, D. A (2013). Fundamentals of Astrodynamics and Applications, Microcosm Press, Hawthorn, CA, USA.